Refactor : 번들 크기 최소화를 위해 아이콘 파일 분리
