+++ src/app/(home)/_components/LandingHeroSection.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import Button from "@/components/ui/Button";
-import { IconCaretDoubleDown } from "@/components/ui/Icons";
+import IconCaretDoubleDown from "@/components/ui/icons/IconCaretDoubleDown";
 import { loginWithGithub } from "@/lib/actions";
 import Link from "next/link";
 import { useTransition } from "react";
@@ -16,9 +16,9 @@ export default function LandingHeroSection({
 
   return (
     <section className="flex-center-center relative min-h-[calc(100dvh-8.5rem)] flex-col items-center text-[3.75rem] leading-[4.538rem] tracking-[0.015em] text-primary-500">
-    <div className="absolute inset-0 -z-10">
-      <EllipsesAnimation />
-    </div>
+      <div className="absolute inset-0 -z-10">
+        <EllipsesAnimation />
+      </div>
       <span className="mb-5">Find your Flaw,</span>
       <h1 className="mb-10 rounded-full border-4 border-primary-500 bg-white px-10 py-[1.156rem]">
         FlawDetector
