+++ src/components/layout/Header.tsx
@@ -2,7 +2,7 @@
 
 import { Aldrich } from "next/font/google";
 import Link from "next/link";
-import { IconBug } from "../ui/Icons";
+import IconBug from "../ui/icons/IconBug";
 import { cn } from "@/lib/utils";
 import { usePathname } from "next/navigation";
 
@@ -33,10 +33,14 @@ function NavMenu({ isDocsPage }: { isDocsPage: boolean }) {
         )}
       >
         <li>
-          <Link prefetch href={"/vuldb/items"}>취약점 DB</Link>
+          <Link prefetch href={"/vuldb/items"}>
+            취약점 DB
+          </Link>
         </li>
         <li>
-          <Link prefetch href={"/repos"}>MY 저장소</Link>
+          <Link prefetch href={"/repos"}>
+            MY 저장소
+          </Link>
         </li>
       </ul>
     </nav>
