+++ src/components/ui/Alert.tsx
@@ -6,13 +6,11 @@ import { useDetectedModeStore } from "@/stores/useDetectedModeStore";
 import { useFileProcessStore } from "@/stores/useFileProcessStore";
 import { useFileViewerStore } from "@/stores/useFileViewerStore";
 import Button from "./Button";
-import {
-  IconArrowsCounterClockwise,
-  IconCheck,
-  IconClose,
-  IconHourGlass,
-  IconThinClose,
-} from "./Icons";
+import IconArrowsCounterClockwise from "./icons/IconArrowsCounterClockwise";
+import IconHourGlass from "./icons/IconHourGlass";
+import IconClose from "./icons/IconClose";
+import IconCheck from "./icons/IconCheck";
+import IconThinClose from "./icons/IconThinClose";
 
 const alertType = {
   onCheck: {
