--- src/components/ui/Icons.tsx
@@ -1,675 +1,0 @@
-import { cn } from "@/lib/utils";
-
-export function IconClose({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="12"
-      height="12"
-      viewBox="0 0 12 12"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn(className)}
-      {...props}
-    >
-      <path d="M7.30975 5.99999L9.78007 3.53061C9.84975 3.46093 9.90502 3.37821 9.94273 3.28716C9.98045 3.19612 9.99986 3.09854 9.99986 2.99999C9.99986 2.90144 9.98045 2.80386 9.94273 2.71282C9.90502 2.62177 9.84975 2.53905 9.78007 2.46936C9.71038 2.39968 9.62766 2.34441 9.53661 2.30669C9.44557 2.26898 9.34799 2.24957 9.24944 2.24957C9.15089 2.24957 9.05331 2.26898 8.96227 2.30669C8.87122 2.34441 8.7885 2.39968 8.71881 2.46936L6.24944 4.93968L3.78007 2.46936C3.63934 2.32863 3.44846 2.24957 3.24944 2.24957C3.05042 2.24957 2.85955 2.32863 2.71882 2.46936C2.57808 2.61009 2.49902 2.80097 2.49902 2.99999C2.49902 3.19901 2.57808 3.38988 2.71882 3.53061L5.18913 5.99999L2.71882 8.46936C2.64913 8.53905 2.59386 8.62177 2.55615 8.71282C2.51843 8.80386 2.49902 8.90144 2.49902 8.99999C2.49902 9.09854 2.51843 9.19612 2.55615 9.28716C2.59386 9.37821 2.64913 9.46093 2.71882 9.53061C2.7885 9.6003 2.87122 9.65557 2.96227 9.69328C3.05331 9.731 3.15089 9.75041 3.24944 9.75041C3.34799 9.75041 3.44557 9.731 3.53661 9.69328C3.62766 9.65557 3.71038 9.6003 3.78007 9.53061L6.24944 7.0603L8.71881 9.53061C8.7885 9.6003 8.87122 9.65557 8.96227 9.69328C9.05331 9.731 9.15089 9.75041 9.24944 9.75041C9.34799 9.75041 9.44557 9.731 9.53661 9.69328C9.62766 9.65557 9.71038 9.6003 9.78007 9.53061C9.84975 9.46093 9.90502 9.37821 9.94273 9.28716C9.98045 9.19612 9.99986 9.09854 9.99986 8.99999C9.99986 8.90144 9.98045 8.80386 9.94273 8.71282C9.90502 8.62177 9.84975 8.53905 9.78007 8.46936L7.30975 5.99999Z" />
-    </svg>
-  );
-}
-
-export function IconThinClose({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="32"
-      height="32"
-      viewBox="0 0 32 32"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn("fill-[#343330]", className)}
-      {...props}
-    >
-      <path d="M25.7081 24.2926C25.801 24.3855 25.8747 24.4958 25.9249 24.6172C25.9752 24.7386 26.0011 24.8687 26.0011 25.0001C26.0011 25.1315 25.9752 25.2616 25.9249 25.383C25.8747 25.5044 25.801 25.6147 25.7081 25.7076C25.6151 25.8005 25.5048 25.8742 25.3835 25.9245C25.2621 25.9747 25.132 26.0006 25.0006 26.0006C24.8692 26.0006 24.7391 25.9747 24.6177 25.9245C24.4963 25.8742 24.386 25.8005 24.2931 25.7076L16.0006 17.4138L7.70806 25.7076C7.52042 25.8952 7.26592 26.0006 7.00056 26.0006C6.73519 26.0006 6.4807 25.8952 6.29306 25.7076C6.10542 25.5199 6 25.2654 6 25.0001C6 24.7347 6.10542 24.4802 6.29306 24.2926L14.5868 16.0001L6.29306 7.70757C6.10542 7.51993 6 7.26543 6 7.00007C6 6.7347 6.10542 6.48021 6.29306 6.29257C6.4807 6.10493 6.73519 5.99951 7.00056 5.99951C7.26592 5.99951 7.52042 6.10493 7.70806 6.29257L16.0006 14.5863L24.2931 6.29257C24.4807 6.10493 24.7352 5.99951 25.0006 5.99951C25.2659 5.99951 25.5204 6.10493 25.7081 6.29257C25.8957 6.48021 26.0011 6.7347 26.0011 7.00007C26.0011 7.26543 25.8957 7.51993 25.7081 7.70757L17.4143 16.0001L25.7081 24.2926Z" />
-    </svg>
-  );
-}
-
-export function IconChat({ className, ...props }: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="35"
-      height="35"
-      viewBox="0 0 35 35"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn(className)}
-      {...props}
-    >
-      <path
-        d="M33.8906 19.6112C32.8406 27.6437 24.8956 34.1537 16.7931 34.1537H5.12059C4.12986 34.1541 3.21173 33.6342 2.70253 32.7843C2.19334 31.9345 2.16792 30.8796 2.63559 30.0062L3.10809 29.1137C3.61559 28.2494 3.61559 27.1781 3.10809 26.3137C-0.86958 20.0304 -0.248309 11.8807 4.63584 6.27293C9.51999 0.665183 17.5073 -1.06902 24.2772 2.00841C31.047 5.08583 34.992 12.2441 33.9781 19.6112H33.8906Z"
-        fill="currentColor"
-      />
-    </svg>
-  );
-}
-
-export function IconArrow({
-  direction = "up",
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  const paths: Record<string, string> = {
-    up: "M26.0612 15.0613C25.9219 15.2011 25.7563 15.3121 25.574 15.3878C25.3917 15.4635 25.1962 15.5025 24.9987 15.5025C24.8013 15.5025 24.6059 15.4635 24.4235 15.3878C24.2412 15.3121 24.0756 15.2011 23.9362 15.0613L17.5 8.62505V27C17.5 27.3979 17.342 27.7794 17.0607 28.0607C16.7794 28.342 16.3978 28.5 16 28.5C15.6022 28.5 15.2206 28.342 14.9393 28.0607C14.658 27.7794 14.5 27.3979 14.5 27V8.62505L8.06125 15.0613C7.77946 15.3431 7.39726 15.5014 6.99875 15.5014C6.60023 15.5014 6.21804 15.3431 5.93625 15.0613C5.65446 14.7795 5.49615 14.3973 5.49615 13.9988C5.49615 13.6003 5.65446 13.2181 5.93625 12.9363L14.9362 3.9363C15.0756 3.79646 15.2412 3.6855 15.4235 3.6098C15.6058 3.53409 15.8013 3.49512 15.9987 3.49512C16.1962 3.49512 16.3916 3.53409 16.574 3.6098C16.7563 3.6855 16.9219 3.79646 17.0612 3.9363L26.0612 12.9363C26.2011 13.0757 26.312 13.2412 26.3878 13.4236C26.4635 13.6059 26.5024 13.8014 26.5024 13.9988C26.5024 14.1962 26.4635 14.3917 26.3878 14.574C26.312 14.7564 26.2011 14.9219 26.0612 15.0613Z",
-    right:
-      "M17.4622 26.5846C17.3224 26.4453 17.2114 26.2797 17.1357 26.0974C17.06 25.915 17.021 25.7196 17.021 25.5221C17.021 25.3247 17.06 25.1292 17.1357 24.9469C17.2114 24.7646 17.3224 24.599 17.4622 24.4596L23.8984 18.0234H5.52344C5.12562 18.0234 4.74409 17.8653 4.46278 17.584C4.18148 17.3027 4.02344 16.9212 4.02344 16.5234C4.02344 16.1256 4.18148 15.744 4.46278 15.4627C4.74409 15.1814 5.12562 15.0234 5.52344 15.0234L23.8984 15.0234L17.4622 8.58463C17.1804 8.30284 17.0221 7.92065 17.0221 7.52213C17.0221 7.12362 17.1804 6.74143 17.4622 6.45963C17.744 6.17784 18.1262 6.01953 18.5247 6.01953C18.9232 6.01953 19.3054 6.17784 19.5872 6.45963L28.5872 15.4596C28.727 15.599 28.838 15.7646 28.9137 15.9469C28.9894 16.1292 29.0284 16.3247 29.0284 16.5221C29.0284 16.7196 28.9894 16.915 28.9137 17.0974C28.838 17.2797 28.727 17.4453 28.5872 17.5846L19.5872 26.5846C19.4478 26.7245 19.2822 26.8354 19.0999 26.9111C18.9176 26.9868 18.7221 27.0258 18.5247 27.0258C18.3273 27.0258 18.1318 26.9868 17.9495 26.9111C17.7671 26.8354 17.6015 26.7245 17.4622 26.5846Z",
-  };
-
-  return (
-    <svg
-      width="32"
-      height="32"
-      viewBox="0 0 32 32"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn("cursor-pointer", className)}
-      {...props}
-    >
-      <path d={paths[direction]} />
-    </svg>
-  );
-}
-
-export function IconFilledPin({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="26"
-      height="26"
-      viewBox="0 0 26 26"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn(className)}
-      {...props}
-    >
-      <path
-        d="M25.4155 8.17095L17.8292 0.585948C17.6435 0.400182 17.423 0.252823 17.1803 0.152285C16.9376 0.0517467 16.6775 0 16.4148 0C16.1522 0 15.892 0.0517467 15.6494 0.152285C15.4067 0.252823 15.1862 0.400182 15.0005 0.585948L8.29671 7.31095C6.96421 6.89345 3.92171 6.3897 0.746712 8.95345C0.527869 9.12942 0.348536 9.34951 0.220388 9.59938C0.0922392 9.84926 0.0181477 10.1233 0.00293644 10.4037C-0.0122749 10.6841 0.0317348 10.9646 0.132101 11.2269C0.232467 11.4891 0.386939 11.7273 0.585462 11.9259L6.62546 17.9634L1.29296 23.2922C1.10532 23.4798 0.999906 23.7343 0.999906 23.9997C0.999906 24.2651 1.10532 24.5196 1.29296 24.7072C1.4806 24.8948 1.7351 25.0003 2.00046 25.0003C2.26583 25.0003 2.52032 24.8948 2.70796 24.7072L8.03671 19.3747L14.073 25.4109C14.2585 25.5973 14.479 25.7452 14.7218 25.8462C14.9646 25.9473 15.225 25.9994 15.488 25.9997C15.5355 25.9997 15.5817 25.9997 15.6292 25.9997C15.9143 25.9799 16.1918 25.8991 16.4429 25.7625C16.6939 25.626 16.9127 25.437 17.0842 25.2084C19.5392 21.9459 19.303 19.2934 18.733 17.7084L25.4167 10.9997C25.6024 10.8139 25.7497 10.5933 25.8501 10.3506C25.9505 10.1079 26.0022 9.84775 26.002 9.58506C26.0019 9.32238 25.9501 9.06229 25.8494 8.81966C25.7488 8.57702 25.6013 8.35659 25.4155 8.17095ZM24.0005 9.58595L16.8417 16.7684C16.6939 16.9168 16.5964 17.1079 16.5629 17.3147C16.5295 17.5215 16.5619 17.7335 16.6555 17.9209C17.838 20.2872 16.4305 22.7447 15.488 23.9984L2.00046 10.5097C3.51046 9.2922 4.95546 8.97095 6.06046 8.97095C6.76162 8.96138 7.457 9.09892 8.10171 9.3747C8.2898 9.46883 8.50279 9.50126 8.71037 9.46737C8.91796 9.43349 9.10957 9.33501 9.25796 9.18595L16.4155 1.9997L24.0005 9.5847V9.58595Z"
-        fill="#6100FF"
-      />
-      <path
-        d="M24.0005 9.58595L16.8417 16.7684C16.6939 16.9168 16.5964 17.1079 16.5629 17.3147C16.5295 17.5215 16.5619 17.7335 16.6555 17.9209C17.838 20.2872 16.4305 22.7447 15.488 23.9984L2.00046 10.5097C3.51046 9.2922 4.95546 8.97095 6.06046 8.97095C6.76162 8.96138 7.457 9.09892 8.10171 9.3747C8.2898 9.46883 8.50279 9.50126 8.71037 9.46737C8.91796 9.43349 9.10957 9.33501 9.25796 9.18595L16.4155 1.9997L24.0005 9.5847V9.58595Z"
-        fill="#6100FF"
-      />
-    </svg>
-  );
-}
-
-export function IconPin({ className, ...props }: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="26"
-      height="26"
-      viewBox="0 0 26 26"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn(className)}
-      {...props}
-    >
-      <path
-        d="M25.4155 8.17095L17.8292 0.585948C17.6435 0.400182 17.423 0.252823 17.1803 0.152285C16.9376 0.0517467 16.6775 0 16.4148 0C16.1522 0 15.892 0.0517467 15.6494 0.152285C15.4067 0.252823 15.1862 0.400182 15.0005 0.585948L8.29671 7.31095C6.96421 6.89345 3.92171 6.3897 0.746712 8.95345C0.527869 9.12942 0.348536 9.34951 0.220388 9.59938C0.0922392 9.84926 0.0181477 10.1233 0.00293644 10.4037C-0.0122749 10.6841 0.0317348 10.9646 0.132101 11.2269C0.232467 11.4891 0.386939 11.7273 0.585462 11.9259L6.62546 17.9634L1.29296 23.2922C1.10532 23.4798 0.999906 23.7343 0.999906 23.9997C0.999906 24.2651 1.10532 24.5196 1.29296 24.7072C1.4806 24.8948 1.7351 25.0003 2.00046 25.0003C2.26583 25.0003 2.52032 24.8948 2.70796 24.7072L8.03671 19.3747L14.073 25.4109C14.2585 25.5973 14.479 25.7452 14.7218 25.8462C14.9646 25.9473 15.225 25.9994 15.488 25.9997C15.5355 25.9997 15.5817 25.9997 15.6292 25.9997C15.9143 25.9799 16.1918 25.8991 16.4429 25.7625C16.6939 25.626 16.9127 25.437 17.0842 25.2084C19.5392 21.9459 19.303 19.2934 18.733 17.7084L25.4167 10.9997C25.6024 10.8139 25.7497 10.5933 25.8501 10.3506C25.9505 10.1079 26.0022 9.84775 26.002 9.58506C26.0019 9.32238 25.9501 9.06229 25.8494 8.81966C25.7488 8.57702 25.6013 8.35659 25.4155 8.17095ZM24.0005 9.58595L16.8417 16.7684C16.6939 16.9168 16.5964 17.1079 16.5629 17.3147C16.5295 17.5215 16.5619 17.7335 16.6555 17.9209C17.838 20.2872 16.4305 22.7447 15.488 23.9984L2.00046 10.5097C3.51046 9.2922 4.95546 8.97095 6.06046 8.97095C6.76162 8.96138 7.457 9.09892 8.10171 9.3747C8.2898 9.46883 8.50279 9.50126 8.71037 9.46737C8.91796 9.43349 9.10957 9.33501 9.25796 9.18595L16.4155 1.9997L24.0005 9.5847V9.58595Z"
-        fill="#6100FF"
-      />
-      <path
-        d="M24.0005 9.58595L16.8417 16.7684C16.6939 16.9168 16.5964 17.1079 16.5629 17.3147C16.5295 17.5215 16.5619 17.7335 16.6555 17.9209C17.838 20.2872 16.4305 22.7447 15.488 23.9984L2.00046 10.5097C3.51046 9.2922 4.95546 8.97095 6.06046 8.97095C6.76162 8.96138 7.457 9.09892 8.10171 9.3747C8.2898 9.46883 8.50279 9.50126 8.71037 9.46737C8.91796 9.43349 9.10957 9.33501 9.25796 9.18595L16.4155 1.9997L24.0005 9.5847V9.58595Z"
-        fill="#FFFFFF"
-      />
-    </svg>
-  );
-}
-
-export function IconFilledShare({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="28"
-      height="23"
-      viewBox="0 0 28 23"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn(className)}
-      {...props}
-    >
-      <path
-        d="M27.7074 10.2933L17.7074 0.293287C17.5676 0.153351 17.3895 0.0580095 17.1955 0.0193193C17.0016 -0.0193709 16.8005 0.000328541 16.6177 0.075926C16.4349 0.151524 16.2787 0.279623 16.1687 0.444026C16.0588 0.608428 16 0.801748 15.9999 0.999537V6.04329C12.7574 6.32079 9.17617 7.90829 6.22992 10.407C2.68242 13.417 0.473674 17.2958 0.00992401 21.3283C-0.0263163 21.6418 0.0372509 21.9587 0.191579 22.234C0.345907 22.5093 0.583132 22.7289 0.869495 22.8616C1.15586 22.9942 1.47677 23.0332 1.78655 22.9729C2.09634 22.9126 2.37921 22.7562 2.59492 22.5258C3.96992 21.062 8.86242 16.4333 15.9999 16.0258V20.9995C16 21.1973 16.0588 21.3906 16.1687 21.555C16.2787 21.7194 16.4349 21.8475 16.6177 21.9231C16.8005 21.9987 17.0016 22.0184 17.1955 21.9798C17.3895 21.9411 17.5676 21.8457 17.7074 21.7058L27.7074 11.7058C27.8944 11.5183 27.9995 11.2643 27.9995 10.9995C27.9995 10.7347 27.8944 10.4808 27.7074 10.2933ZM17.9999 18.5858V14.9995C17.9999 14.7343 17.8946 14.48 17.707 14.2924C17.5195 14.1049 17.2651 13.9995 16.9999 13.9995C13.4899 13.9995 10.0712 14.9158 6.83867 16.7245C5.19235 17.6498 3.65844 18.7622 2.26742 20.0395C2.99242 17.0595 4.81992 14.2258 7.52367 11.932C10.4262 9.47079 13.9687 7.99954 16.9999 7.99954C17.2651 7.99954 17.5195 7.89418 17.707 7.70664C17.8946 7.51911 17.9999 7.26475 17.9999 6.99954V3.41454L25.5862 10.9995L17.9999 18.5858Z"
-        fill="#6100FF"
-      />
-      <path
-        d="M17.9999 18.5858V14.9995C17.9999 14.7343 17.8946 14.48 17.707 14.2924C17.5195 14.1049 17.2651 13.9995 16.9999 13.9995C13.4899 13.9995 10.0712 14.9158 6.83867 16.7245C5.19235 17.6498 3.65844 18.7622 2.26742 20.0395C2.99242 17.0595 4.81992 14.2258 7.52367 11.932C10.4262 9.47079 13.9687 7.99954 16.9999 7.99954C17.2651 7.99954 17.5195 7.89418 17.707 7.70664C17.8946 7.51911 17.9999 7.26475 17.9999 6.99954V3.41454L25.5862 10.9995L17.9999 18.5858Z"
-        fill="#6100FF"
-      />
-    </svg>
-  );
-}
-
-export function IconShare({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="28"
-      height="23"
-      viewBox="0 0 28 23"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn(className)}
-      {...props}
-    >
-      <path
-        d="M27.7074 10.2933L17.7074 0.293287C17.5676 0.153351 17.3895 0.0580095 17.1955 0.0193193C17.0016 -0.0193709 16.8005 0.000328541 16.6177 0.075926C16.4349 0.151524 16.2787 0.279623 16.1687 0.444026C16.0588 0.608428 16 0.801748 15.9999 0.999537V6.04329C12.7574 6.32079 9.17617 7.90829 6.22992 10.407C2.68242 13.417 0.473674 17.2958 0.00992401 21.3283C-0.0263163 21.6418 0.0372509 21.9587 0.191579 22.234C0.345907 22.5093 0.583132 22.7289 0.869495 22.8616C1.15586 22.9942 1.47677 23.0332 1.78655 22.9729C2.09634 22.9126 2.37921 22.7562 2.59492 22.5258C3.96992 21.062 8.86242 16.4333 15.9999 16.0258V20.9995C16 21.1973 16.0588 21.3906 16.1687 21.555C16.2787 21.7194 16.4349 21.8475 16.6177 21.9231C16.8005 21.9987 17.0016 22.0184 17.1955 21.9798C17.3895 21.9411 17.5676 21.8457 17.7074 21.7058L27.7074 11.7058C27.8944 11.5183 27.9995 11.2643 27.9995 10.9995C27.9995 10.7347 27.8944 10.4808 27.7074 10.2933ZM17.9999 18.5858V14.9995C17.9999 14.7343 17.8946 14.48 17.707 14.2924C17.5195 14.1049 17.2651 13.9995 16.9999 13.9995C13.4899 13.9995 10.0712 14.9158 6.83867 16.7245C5.19235 17.6498 3.65844 18.7622 2.26742 20.0395C2.99242 17.0595 4.81992 14.2258 7.52367 11.932C10.4262 9.47079 13.9687 7.99954 16.9999 7.99954C17.2651 7.99954 17.5195 7.89418 17.707 7.70664C17.8946 7.51911 17.9999 7.26475 17.9999 6.99954V3.41454L25.5862 10.9995L17.9999 18.5858Z"
-        fill="#6100FF"
-      />
-      <path
-        d="M17.9999 18.5858V14.9995C17.9999 14.7343 17.8946 14.48 17.707 14.2924C17.5195 14.1049 17.2651 13.9995 16.9999 13.9995C13.4899 13.9995 10.0712 14.9158 6.83867 16.7245C5.19235 17.6498 3.65844 18.7622 2.26742 20.0395C2.99242 17.0595 4.81992 14.2258 7.52367 11.932C10.4262 9.47079 13.9687 7.99954 16.9999 7.99954C17.2651 7.99954 17.5195 7.89418 17.707 7.70664C17.8946 7.51911 17.9999 7.26475 17.9999 6.99954V3.41454L25.5862 10.9995L17.9999 18.5858Z"
-        fill="#FFFFFF"
-      />
-    </svg>
-  );
-}
-
-export function IconFolder({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn(className)}
-      {...props}
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M19.45 7.625V7.905C19.9453 8.52698 20.146 9.33343 20 10.115L17.92 16.615C17.5115 17.6736 16.4947 18.3727 15.36 18.375H2.7C1.21076 18.3427 0.015841 17.1345 1.51304e-08 15.645V4.375C-0.000182366 2.3233 1.64848 0.652356 3.7 0.625H8C9.38071 0.625 10.5 1.74429 10.5 3.125C10.5 3.67728 10.9477 4.125 11.5 4.125H16C17.9134 4.15234 19.4502 5.71136 19.45 7.625ZM3.7 2.125C2.45736 2.125 1.45 3.13236 1.45 4.375V11.495L2.52 8.715C2.91802 7.63378 3.94784 6.91529 5.1 6.915H17.35C17.523 6.89995 17.697 6.89995 17.87 6.915C17.5751 6.13835 16.8308 5.62489 16 5.625H11.5C10.1216 5.61952 9.00548 4.50344 9 3.125C9 2.57272 8.55229 2.125 8 2.125H3.7Z"
-      />
-    </svg>
-  );
-}
-
-export function IconDoc({ className, ...props }: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn(className)}
-      {...props}
-    >
-      <path
-        d="M20.0306 7.71938L14.7806 2.46938C14.7109 2.39975 14.6282 2.34454 14.5371 2.3069C14.4461 2.26926 14.3485 2.24992 14.25 2.25H5.25C4.85218 2.25 4.47064 2.40804 4.18934 2.68934C3.90804 2.97064 3.75 3.35218 3.75 3.75V20.25C3.75 20.6478 3.90804 21.0294 4.18934 21.3107C4.47064 21.592 4.85218 21.75 5.25 21.75H18.75C19.1478 21.75 19.5294 21.592 19.8107 21.3107C20.092 21.0294 20.25 20.6478 20.25 20.25V8.25C20.2501 8.15148 20.2307 8.05391 20.1931 7.96286C20.1555 7.87182 20.1003 7.78908 20.0306 7.71938ZM15 4.81031L17.6897 7.5H15V4.81031ZM18.75 20.25H5.25V3.75H13.5V8.25C13.5 8.44891 13.579 8.63968 13.7197 8.78033C13.8603 8.92098 14.0511 9 14.25 9H18.75V20.25Z"
-        fill="#969696"
-      />
-    </svg>
-  );
-}
-
-export function IconDone({ className, ...props }: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="20"
-      height="20"
-      viewBox="0 0 20 20"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn(className)}
-      {...props}
-    >
-      <path d="M10 1.875C8.39303 1.875 6.82214 2.35152 5.486 3.24431C4.14985 4.1371 3.10844 5.40605 2.49348 6.8907C1.87852 8.37535 1.71762 10.009 2.03112 11.5851C2.34463 13.1612 3.11846 14.6089 4.25476 15.7452C5.39106 16.8815 6.8388 17.6554 8.4149 17.9689C9.99099 18.2824 11.6247 18.1215 13.1093 17.5065C14.594 16.8916 15.8629 15.8502 16.7557 14.514C17.6485 13.1779 18.125 11.607 18.125 10C18.1227 7.84581 17.266 5.78051 15.7427 4.25727C14.2195 2.73403 12.1542 1.87727 10 1.875ZM13.5672 8.56719L9.19219 12.9422C9.13415 13.0003 9.06522 13.0464 8.98934 13.0779C8.91347 13.1093 8.83214 13.1255 8.75 13.1255C8.66787 13.1255 8.58654 13.1093 8.51067 13.0779C8.43479 13.0464 8.36586 13.0003 8.30782 12.9422L6.43282 11.0672C6.31554 10.9499 6.24966 10.7909 6.24966 10.625C6.24966 10.4591 6.31554 10.3001 6.43282 10.1828C6.55009 10.0655 6.70915 9.99965 6.875 9.99965C7.04086 9.99965 7.19992 10.0655 7.31719 10.1828L8.75 11.6164L12.6828 7.68281C12.7409 7.62474 12.8098 7.57868 12.8857 7.54725C12.9616 7.51583 13.0429 7.49965 13.125 7.49965C13.2071 7.49965 13.2884 7.51583 13.3643 7.54725C13.4402 7.57868 13.5091 7.62474 13.5672 7.68281C13.6253 7.74088 13.6713 7.80982 13.7027 7.88569C13.7342 7.96156 13.7504 8.04288 13.7504 8.125C13.7504 8.20712 13.7342 8.28844 13.7027 8.36431C13.6713 8.44018 13.6253 8.50912 13.5672 8.56719Z" />
-    </svg>
-  );
-}
-
-export function IconOnProcess({
-  className,
-  size = 20,
-  ...props
-}: React.ComponentProps<"svg"> & { size?: number }) {
-  return (
-    <svg
-      width={size}
-      height={size}
-      viewBox="0 0 40 40"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn(className, "animate-spin")}
-      {...props}
-    >
-      <rect x="0.5" width="40" height="40" rx="20" fill="#6100FF" />
-      <path
-        d="M27.2959 25.2041C27.4008 25.3086 27.484 25.4328 27.5408 25.5695C27.5976 25.7063 27.6268 25.8529 27.6268 26.0009C27.6268 26.149 27.5976 26.2956 27.5408 26.4324C27.484 26.5691 27.4008 26.6933 27.2959 26.7978C27.19 26.9028 24.6709 29.3769 20.5 29.3769C17.3397 29.3769 14.9359 27.74 13.375 26.2025V27.5C13.375 27.7984 13.2565 28.0845 13.0455 28.2955C12.8345 28.5065 12.5484 28.625 12.25 28.625C11.9516 28.625 11.6655 28.5065 11.4545 28.2955C11.2435 28.0845 11.125 27.7984 11.125 27.5V23C11.125 22.7016 11.2435 22.4155 11.4545 22.2045C11.6655 21.9935 11.9516 21.875 12.25 21.875H16.75C17.0484 21.875 17.3345 21.9935 17.5455 22.2045C17.7565 22.4155 17.875 22.7016 17.875 23C17.875 23.2984 17.7565 23.5845 17.5455 23.7955C17.3345 24.0065 17.0484 24.125 16.75 24.125H14.5113C15.7394 25.4675 17.8 27.125 20.5 27.125C23.7466 27.125 25.6853 25.2237 25.7041 25.2041C25.8085 25.0995 25.9326 25.0165 26.0692 24.9599C26.2058 24.9033 26.3522 24.8741 26.5 24.8741C26.6478 24.8741 26.7942 24.9033 26.9308 24.9599C27.0674 25.0165 27.1915 25.0995 27.2959 25.2041ZM28.75 11.375C28.4516 11.375 28.1655 11.4935 27.9545 11.7045C27.7435 11.9155 27.625 12.2016 27.625 12.5V13.7994C26.0641 12.2619 23.6603 10.625 20.5 10.625C16.3319 10.625 13.81 13.0991 13.7041 13.2041C13.4927 13.4154 13.374 13.7021 13.374 14.0009C13.374 14.2998 13.4927 14.5865 13.7041 14.7978C13.9154 15.0092 14.2021 15.1279 14.5009 15.1279C14.7998 15.1279 15.0865 15.0092 15.2978 14.7978C15.3147 14.7763 17.2534 12.875 20.5 12.875C23.2 12.875 25.2606 14.5325 26.4888 15.875H24.25C23.9516 15.875 23.6655 15.9935 23.4545 16.2045C23.2435 16.4155 23.125 16.7016 23.125 17C23.125 17.2984 23.2435 17.5845 23.4545 17.7955C23.6655 18.0065 23.9516 18.125 24.25 18.125H28.75C29.0484 18.125 29.3345 18.0065 29.5455 17.7955C29.7565 17.5845 29.875 17.2984 29.875 17V12.5C29.875 12.2016 29.7565 11.9155 29.5455 11.7045C29.3345 11.4935 29.0484 11.375 28.75 11.375Z"
-        fill="white"
-      />
-    </svg>
-  );
-}
-
-export function IconError({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="20"
-      height="20"
-      viewBox="0 0 20 20"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn(className)}
-      {...props}
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M17.3796 12.9196L12.6796 4.21959C12.0926 3.15854 10.9756 2.5 9.76298 2.5C8.55035 2.5 7.43337 3.15854 6.84631 4.21959L2.07131 12.9113C1.51031 13.9435 1.53416 15.1949 2.1341 16.2051C2.73403 17.2152 3.82144 17.8349 4.99631 17.8363H14.4463C15.6202 17.8363 16.7075 17.2189 17.3089 16.2108C17.9104 15.2027 17.9372 13.9526 17.3796 12.9196ZM9.04631 8.11959C9.04631 7.77441 9.32613 7.49459 9.67131 7.49459C10.0165 7.49459 10.2963 7.77441 10.2963 8.11959V10.6946C10.2963 11.0398 10.0165 11.3196 9.67131 11.3196C9.32613 11.3196 9.04631 11.0398 9.04631 10.6946V8.11959ZM9.67964 13.7446C10.0248 13.7446 10.3046 13.4648 10.3046 13.1196L10.2963 13.1113C10.2963 12.7707 10.0202 12.4946 9.67964 12.4946C9.33634 12.4991 9.05913 12.7763 9.05464 13.1196C9.05464 13.4648 9.33447 13.7446 9.67964 13.7446Z"
-        fill="#FF6D6D"
-      />
-    </svg>
-  );
-}
-
-export function IconOnWait({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="20"
-      height="20"
-      viewBox="0 0 20 20"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn(className)}
-      {...props}
-    >
-      <path
-        d="M10 5.00008V10.0001L13.3333 11.6667M18.3333 10.0001C18.3333 14.6025 14.6024 18.3334 10 18.3334C5.39763 18.3334 1.66667 14.6025 1.66667 10.0001C1.66667 5.39771 5.39763 1.66675 10 1.66675C14.6024 1.66675 18.3333 5.39771 18.3333 10.0001Z"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      />
-    </svg>
-  );
-}
-
-export function IconCaretLeft({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="32"
-      height="32"
-      viewBox="0 0 32 32"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn("fill-primary-500", className)}
-      {...props}
-    >
-      <path d="M20.7073 25.2926C20.8002 25.3855 20.8739 25.4958 20.9242 25.6172C20.9745 25.7386 21.0004 25.8687 21.0004 26.0001C21.0004 26.1315 20.9745 26.2616 20.9242 26.383C20.8739 26.5044 20.8002 26.6147 20.7073 26.7076C20.6144 26.8005 20.5041 26.8742 20.3827 26.9245C20.2613 26.9747 20.1312 27.0006 19.9998 27.0006C19.8684 27.0006 19.7383 26.9747 19.6169 26.9245C19.4955 26.8742 19.3852 26.8005 19.2923 26.7076L9.29231 16.7076C9.19933 16.6147 9.12557 16.5044 9.07525 16.383C9.02493 16.2616 8.99902 16.1315 8.99902 16.0001C8.99902 15.8687 9.02493 15.7385 9.07525 15.6171C9.12557 15.4957 9.19933 15.3854 9.29231 15.2926L19.2923 5.29257C19.48 5.10493 19.7344 4.99951 19.9998 4.99951C20.2652 4.99951 20.5197 5.10493 20.7073 5.29257C20.895 5.48021 21.0004 5.7347 21.0004 6.00007C21.0004 6.26543 20.895 6.51993 20.7073 6.70757L11.4136 16.0001L20.7073 25.2926Z" />
-    </svg>
-  );
-}
-
-export function IconCaretDown({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="17"
-      height="9"
-      viewBox="0 0 17 9"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn("fill-gray-dark", className)}
-      {...props}
-    >
-      <path d="M7.81225 8.34845L0.821839 1.72595C0.165583 1.10424 0.605592 0 1.50958 0L15.4904 0C16.3944 0 16.8344 1.10424 16.1782 1.72595L9.18774 8.34845C8.80205 8.71385 8.19795 8.71385 7.81225 8.34845Z" />
-    </svg>
-  );
-}
-
-export function IconCheck({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn("fill-[#343330]", className)}
-      {...props}
-    >
-      <path d="M21.7959 7.54597L9.7959 19.546C9.69138 19.6509 9.56719 19.7341 9.43044 19.7909C9.2937 19.8476 9.14709 19.8769 8.99902 19.8769C8.85096 19.8769 8.70435 19.8476 8.5676 19.7909C8.43086 19.7341 8.30666 19.6509 8.20215 19.546L2.95215 14.296C2.8475 14.1913 2.76449 14.0671 2.70785 13.9304C2.65122 13.7936 2.62207 13.6471 2.62207 13.4991C2.62207 13.3511 2.65122 13.2046 2.70785 13.0678C2.76449 12.9311 2.8475 12.8069 2.95215 12.7022C3.05679 12.5976 3.18103 12.5146 3.31776 12.4579C3.45448 12.4013 3.60103 12.3721 3.74902 12.3721C3.89702 12.3721 4.04356 12.4013 4.18029 12.4579C4.31702 12.5146 4.44125 12.5976 4.5459 12.7022L8.99996 17.1563L20.204 5.9541C20.4154 5.74276 20.702 5.62402 21.0009 5.62402C21.2998 5.62402 21.5864 5.74276 21.7978 5.9541C22.0091 6.16544 22.1278 6.45209 22.1278 6.75098C22.1278 7.04986 22.0091 7.33651 21.7978 7.54785L21.7959 7.54597Z" />
-    </svg>
-  );
-}
-
-export function IconTriangle({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="23"
-      height="20"
-      viewBox="0 0 23 20"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn("fill-accent-orange", className)}
-      {...props}
-    >
-      <path d="M9.80419 0.958085C10.5579 -0.319361 12.4421 -0.319362 13.1958 0.958083L22.7347 17.1257C23.4884 18.4032 22.5463 20 21.0389 20H1.96108C0.453689 20 -0.488429 18.4032 0.265264 17.1257L9.80419 0.958085Z" />
-    </svg>
-  );
-}
-
-export function IconCircle({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="22"
-      height="22"
-      viewBox="0 0 22 22"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn("fill-accent-green", className)}
-      {...props}
-    >
-      <circle cx="11" cy="11" r="11" />
-    </svg>
-  );
-}
-
-export function IconBug({
-  className,
-  width = "191",
-  height = "196",
-  color = "fill-primary-500",
-  ...props
-}: React.ComponentProps<"svg"> & { width?: string; height?: string }) {
-  return (
-    <svg
-      width={width}
-      height={height}
-      viewBox="0 0 191 196"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn(color, className)}
-      {...props}
-    >
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M95.3023 20.0367C69.6385 20.0367 50.4865 38.8569 50.4865 60.1101C50.4865 65.6431 46.0277 70.1284 40.5274 70.1284C35.0272 70.1284 30.5684 65.6431 30.5684 60.1101C30.5684 26.0366 60.467 0 95.3023 0C130.138 0 160.036 26.0366 160.036 60.1101C160.036 65.6431 155.577 70.1284 150.077 70.1284C144.577 70.1284 140.118 65.6431 140.118 60.1101C140.118 38.8569 120.966 20.0367 95.3023 20.0367Z"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M40.5274 120.477C46.0277 120.477 50.4865 124.962 50.4865 130.495C50.4865 151.749 69.6385 170.569 95.3023 170.569C120.966 170.569 140.118 151.749 140.118 130.495C140.118 124.962 144.577 120.477 150.077 120.477C155.577 120.477 160.036 124.962 160.036 130.495C160.036 164.569 130.138 190.605 95.3023 190.605C60.467 190.605 30.5684 164.569 30.5684 130.495C30.5684 124.962 35.0272 120.477 40.5274 120.477Z"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M0 96.2024C0 90.7404 4.49142 86.3125 10.0319 86.3125H180.574C186.114 86.3125 190.606 90.7404 190.606 96.2024C190.606 101.664 186.114 106.092 180.574 106.092H10.0319C4.49142 106.092 0 101.664 0 96.2024Z"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M96.2024 131.267C101.664 131.267 106.092 135.658 106.092 141.075V175.404C106.092 180.821 101.664 185.212 96.2024 185.212C90.7404 185.212 86.3125 180.821 86.3125 175.404V141.075C86.3125 135.658 90.7404 131.267 96.2024 131.267Z"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M33.4274 3.00955C37.2395 -1.00318 43.4201 -1.00318 47.2322 3.00955L61.8744 18.4224C65.6865 22.4351 65.6865 28.941 61.8744 32.9538C58.0623 36.9665 51.8817 36.9665 48.0696 32.9538L33.4274 17.5409C29.6153 13.5282 29.6153 7.02227 33.4274 3.00955Z"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M157.178 3.00955C160.99 7.02227 160.99 13.5282 157.178 17.5409L142.536 32.9538C138.724 36.9665 132.543 36.9665 128.731 32.9538C124.919 28.941 124.919 22.4351 128.731 18.4224L143.373 3.00955C147.185 -1.00318 153.366 -1.00318 157.178 3.00955Z"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M61.8744 163.047C65.6865 167.059 65.6865 173.565 61.8744 177.578L47.2322 192.991C43.4201 197.004 37.2395 197.004 33.4274 192.991C29.6153 188.978 29.6153 182.472 33.4274 178.459L48.0696 163.047C51.8817 159.034 58.0623 159.034 61.8744 163.047Z"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M128.731 163.047C132.543 159.034 138.724 159.034 142.536 163.047L157.178 178.459C160.99 182.472 160.99 188.978 157.178 192.991C153.366 197.004 147.185 197.004 143.373 192.991L128.731 177.578C124.919 173.565 124.919 167.059 128.731 163.047Z"
-      />
-    </svg>
-  );
-}
-
-export function IconMagnifierWithPlus({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="27"
-      height="27"
-      viewBox="0 0 27 27"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn("fill-primary-500", className)}
-      {...props}
-    >
-      <path
-        d="M26.2078 25.2923L19.949 19.0348C21.7631 16.8569 22.6676 14.0635 22.4746 11.2357C22.2815 8.40789 21.0057 5.76337 18.9125 3.85226C16.8193 1.94116 14.0698 0.910614 11.2362 0.975014C8.4025 1.03941 5.70274 2.1938 3.69851 4.19802C1.69429 6.20225 0.539902 8.90201 0.475503 11.7357C0.411103 14.5694 1.44164 17.3188 3.35275 19.412C5.26385 21.5052 7.90838 22.781 10.7362 22.9741C13.564 23.1672 16.3574 22.2626 18.5353 20.4485L24.7928 26.7073C24.8857 26.8002 24.996 26.8739 25.1174 26.9242C25.2388 26.9745 25.3689 27.0004 25.5003 27.0004C25.6317 27.0004 25.7618 26.9745 25.8832 26.9242C26.0046 26.8739 26.1149 26.8002 26.2078 26.7073C26.3007 26.6144 26.3744 26.5041 26.4247 26.3827C26.475 26.2613 26.5008 26.1312 26.5008 25.9998C26.5008 25.8684 26.475 25.7383 26.4247 25.6169C26.3744 25.4955 26.3007 25.3852 26.2078 25.2923ZM2.50029 11.9998C2.50029 10.2198 3.02813 8.47971 4.01706 6.99966C5.00599 5.51962 6.4116 4.36607 8.05613 3.68488C9.70067 3.00369 11.5103 2.82546 13.2561 3.17273C15.0019 3.52 16.6056 4.37716 17.8642 5.63584C19.1229 6.89451 19.9801 8.49816 20.3274 10.244C20.6746 11.9898 20.4964 13.7994 19.8152 15.4439C19.134 17.0885 17.9805 18.4941 16.5004 19.483C15.0204 20.472 13.2803 20.9998 11.5003 20.9998C9.11415 20.9972 6.8265 20.0481 5.13925 18.3608C3.45199 16.6736 2.50293 14.3859 2.50029 11.9998Z"
-        fill="#6100FF"
-      />
-    </svg>
-  );
-}
-
-export function IconCaretDoubleDown({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="56"
-      height="57"
-      viewBox="0 0 56 57"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn("fill-primary-500", className)}
-      {...props}
-    >
-      <path
-        d="M46.7381 27.2619C46.9008 27.4244 47.0299 27.6174 47.118 27.8299C47.206 28.0423 47.2514 28.27 47.2514 28.5C47.2514 28.73 47.206 28.9577 47.118 29.1701C47.0299 29.3826 46.9008 29.5756 46.7381 29.7381L29.2381 47.2381C29.0756 47.4008 28.8826 47.5299 28.6701 47.618C28.4577 47.706 28.23 47.7514 28 47.7514C27.77 47.7514 27.5423 47.706 27.3299 47.618C27.1174 47.5299 26.9244 47.4008 26.7619 47.2381L9.26187 29.7381C8.9335 29.4098 8.74902 28.9644 8.74902 28.5C8.74902 28.0356 8.9335 27.5902 9.26187 27.2619C9.59024 26.9335 10.0356 26.749 10.5 26.749C10.9644 26.749 11.4098 26.9335 11.7381 27.2619L28 43.5259L44.2619 27.2619C44.4244 27.0992 44.6174 26.9701 44.8298 26.882C45.0423 26.7939 45.27 26.7486 45.5 26.7486C45.73 26.7486 45.9577 26.7939 46.1701 26.882C46.3826 26.9701 46.5756 27.0992 46.7381 27.2619ZM26.7619 29.7381C26.9244 29.9008 27.1174 30.0299 27.3299 30.118C27.5423 30.206 27.77 30.2514 28 30.2514C28.23 30.2514 28.4577 30.206 28.6701 30.118C28.8826 30.0299 29.0756 29.9008 29.2381 29.7381L46.7381 12.2381C47.0665 11.9098 47.251 11.4644 47.251 11C47.251 10.5356 47.0665 10.0902 46.7381 9.76187C46.4098 9.4335 45.9644 9.24902 45.5 9.24902C45.0356 9.24902 44.5902 9.4335 44.2619 9.76187L28 26.0259L11.7381 9.76187C11.4098 9.4335 10.9644 9.24902 10.5 9.24902C10.0356 9.24902 9.59024 9.4335 9.26187 9.76187C8.9335 10.0902 8.74902 10.5356 8.74902 11C8.74902 11.4644 8.9335 11.9098 9.26187 12.2381L26.7619 29.7381Z"
-        fill="#6100FF"
-      />
-    </svg>
-  );
-}
-
-export function IconStar({
-  className,
-  filled = false,
-  ...props
-}: React.ComponentProps<"svg"> & { filled?: boolean }) {
-  return (
-    <svg
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      className={cn(
-        "transition-colors",
-        filled ? "fill-current" : "fill-none stroke-current",
-        className,
-      )}
-      strokeWidth={filled ? "0" : "1.5"}
-      xmlns="http://www.w3.org/2000/svg"
-      {...props}
-    >
-      <path d="M22.4258 9.12124C22.335 8.84129 22.1638 8.59424 21.9336 8.41087C21.7034 8.2275 21.4243 8.11591 21.1311 8.08999L15.5633 7.60999L13.3808 2.41999C13.2672 2.14766 13.0755 1.91504 12.8299 1.75142C12.5844 1.5878 12.2959 1.50049 12.0008 1.50049C11.7057 1.50049 11.4172 1.5878 11.1717 1.75142C10.9261 1.91504 10.7344 2.14766 10.6208 2.41999L8.44486 7.60999L2.87049 8.0928C2.57615 8.11754 2.29566 8.22862 2.0642 8.41213C1.83274 8.59564 1.66061 8.8434 1.5694 9.12434C1.47819 9.40528 1.47195 9.7069 1.55148 9.99138C1.631 10.2758 1.79275 10.5305 2.01643 10.7234L6.24549 14.4191L4.97799 19.9072C4.91097 20.1944 4.93009 20.495 5.03297 20.7713C5.13585 21.0477 5.31792 21.2877 5.55642 21.4611C5.79492 21.6346 6.07926 21.7339 6.3739 21.7467C6.66854 21.7594 6.9604 21.685 7.21299 21.5328L11.9942 18.6266L16.7858 21.5328C17.0384 21.685 17.3303 21.7594 17.6249 21.7467C17.9195 21.7339 18.2039 21.6346 18.4424 21.4611C18.6809 21.2877 18.8629 21.0477 18.9658 20.7713C19.0687 20.495 19.0878 20.1944 19.0208 19.9072L17.7542 14.4134L21.9824 10.7234C22.206 10.5298 22.3674 10.2745 22.4463 9.98938C22.5252 9.70431 22.5181 9.40227 22.4258 9.12124Z" />
-    </svg>
-  );
-}
-
-export function IconArrowsCounterClockwise({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="48"
-      height="48"
-      viewBox="0 0 48 48"
-      fill="=currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn("fill-primary-500", className)}
-      {...props}
-    >
-      <path d="M16.5 19.5H7.5C7.10218 19.5 6.72064 19.342 6.43934 19.0607C6.15804 18.7794 6 18.3978 6 18V9C6 8.60218 6.15804 8.22064 6.43934 7.93934C6.72064 7.65804 7.10218 7.5 7.5 7.5C7.89782 7.5 8.27936 7.65804 8.56066 7.93934C8.84196 8.22064 9 8.60218 9 9V13.4719C11.9006 10.2 16.9894 6 24 6C32.0287 6 36.8588 10.7381 37.0613 10.9388C37.3432 11.2197 37.502 11.6012 37.5027 11.9992C37.5034 12.3973 37.346 12.7793 37.065 13.0612C36.784 13.3432 36.4026 13.502 36.0045 13.5027C35.6065 13.5034 35.2245 13.346 34.9425 13.065C34.8919 13.0144 30.735 9 24 9C17.4375 9 12.7069 13.6219 10.395 16.5H16.5C16.8978 16.5 17.2794 16.658 17.5607 16.9393C17.842 17.2206 18 17.6022 18 18C18 18.3978 17.842 18.7794 17.5607 19.0607C17.2794 19.342 16.8978 19.5 16.5 19.5ZM40.5 28.5H31.5C31.1022 28.5 30.7206 28.658 30.4393 28.9393C30.158 29.2206 30 29.6022 30 30C30 30.3978 30.158 30.7794 30.4393 31.0607C30.7206 31.342 31.1022 31.5 31.5 31.5H37.605C35.2931 34.3781 30.5625 39 24 39C17.265 39 13.1081 34.9856 13.0575 34.935C12.7755 34.654 12.3935 34.4966 11.9955 34.4973C11.5974 34.498 11.216 34.6568 10.935 34.9387C10.654 35.2207 10.4966 35.6027 10.4973 36.0008C10.498 36.3988 10.6568 36.7803 10.9388 37.0613C11.1413 37.2619 15.9713 42 24 42C31.0106 42 36.0994 37.8 39 34.5281V39C39 39.3978 39.158 39.7794 39.4393 40.0607C39.7206 40.342 40.1022 40.5 40.5 40.5C40.8978 40.5 41.2794 40.342 41.5607 40.0607C41.842 39.7794 42 39.3978 42 39V30C42 29.6022 41.842 29.2206 41.5607 28.9393C41.2794 28.658 40.8978 28.5 40.5 28.5Z" />
-    </svg>
-  );
-}
-
-export function IconHourGlass({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="48"
-      height="48"
-      viewBox="0 0 48 48"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn("fill-gray-dark", className)}
-      {...props}
-    >
-      <path d="M37.5 14.1825V7.5C37.5 6.70435 37.1839 5.94129 36.6213 5.37868C36.0587 4.81607 35.2956 4.5 34.5 4.5H13.5C12.7044 4.5 11.9413 4.81607 11.3787 5.37868C10.8161 5.94129 10.5 6.70435 10.5 7.5V14.25C10.501 14.7156 10.6099 15.1745 10.8181 15.591C11.0263 16.0074 11.3282 16.3699 11.7 16.65L21.5006 24L11.7 31.35C11.3282 31.6301 11.0263 31.9926 10.8181 32.409C10.6099 32.8255 10.501 33.2844 10.5 33.75V40.5C10.5 41.2956 10.8161 42.0587 11.3787 42.6213C11.9413 43.1839 12.7044 43.5 13.5 43.5H34.5C35.2956 43.5 36.0587 43.1839 36.6213 42.6213C37.1839 42.0587 37.5 41.2956 37.5 40.5V33.8175C37.4989 33.3537 37.391 32.8965 37.1845 32.4812C36.978 32.0659 36.6785 31.7039 36.3094 31.4231L26.4881 24L36.3094 16.5769C36.6785 16.2961 36.978 15.9341 37.1845 15.5188C37.391 15.1035 37.4989 14.6463 37.5 14.1825ZM34.5 7.5V12H13.5V7.5H34.5ZM34.5 40.5H13.5V33.75L24 25.875L34.5 33.8156V40.5Z" />
-    </svg>
-  );
-}
-
-export function IconCopy({ className, ...props }: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn("fill-gray-light", className)}
-      {...props}
-    >
-      <path d="M20.25 3H8.25C8.05109 3 7.86032 3.07902 7.71967 3.21967C7.57902 3.36032 7.5 3.55109 7.5 3.75V7.5H3.75C3.55109 7.5 3.36032 7.57902 3.21967 7.71967C3.07902 7.86032 3 8.05109 3 8.25V20.25C3 20.4489 3.07902 20.6397 3.21967 20.7803C3.36032 20.921 3.55109 21 3.75 21H15.75C15.9489 21 16.1397 20.921 16.2803 20.7803C16.421 20.6397 16.5 20.4489 16.5 20.25V16.5H20.25C20.4489 16.5 20.6397 16.421 20.7803 16.2803C20.921 16.1397 21 15.9489 21 15.75V3.75C21 3.55109 20.921 3.36032 20.7803 3.21967C20.6397 3.07902 20.4489 3 20.25 3ZM15 19.5H4.5V9H15V19.5ZM19.5 15H16.5V8.25C16.5 8.05109 16.421 7.86032 16.2803 7.71967C16.1397 7.57902 15.9489 7.5 15.75 7.5H9V4.5H19.5V15Z" />
-    </svg>
-  );
-}
-
-export function IconList({ className, ...props }: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={className}
-      {...props}
-    >
-      <path
-        d="M15.37 10.5H5.75C5.33579 10.5 5 10.1642 5 9.75C5 9.33579 5.33579 9 5.75 9H15.37C15.7842 9 16.12 9.33579 16.12 9.75C16.12 10.1642 15.7842 10.5 15.37 10.5Z"
-        fill="black"
-      />
-      <path
-        d="M5.75 14H18.57C18.9842 14 19.32 14.3358 19.32 14.75C19.32 15.1642 18.9842 15.5 18.57 15.5H5.75C5.33579 15.5 5 15.1642 5 14.75C5 14.3358 5.33579 14 5.75 14Z"
-        fill="black"
-      />
-    </svg>
-  );
-}
-
-export function IconClockCounterClockwise({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="29"
-      height="29"
-      viewBox="0 0 29 29"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn("fill-[#343330]", className)}
-      {...props}
-    >
-      <path d="M15.625 9.1807V13.9352L19.5756 16.3054C19.7746 16.4249 19.918 16.6186 19.9742 16.8438C20.0304 17.069 19.9948 17.3073 19.8753 17.5063C19.7558 17.7053 19.5621 17.8487 19.3369 17.9049C19.1117 17.9611 18.8734 17.9255 18.6744 17.806L14.2994 15.181C14.1699 15.1032 14.0627 14.9932 13.9884 14.8617C13.914 14.7303 13.875 14.5818 13.875 14.4307V9.1807C13.875 8.94863 13.9672 8.72607 14.1313 8.56198C14.2954 8.39788 14.5179 8.3057 14.75 8.3057C14.9821 8.3057 15.2046 8.39788 15.3687 8.56198C15.5328 8.72607 15.625 8.94863 15.625 9.1807ZM14.75 3.9307C13.3697 3.92726 12.0024 4.19761 10.7272 4.7261C9.45211 5.25459 8.29444 6.03073 7.32125 7.0096C6.52609 7.8146 5.81953 8.58898 5.125 9.39945V7.4307C5.125 7.19863 5.03281 6.97607 4.86872 6.81198C4.70462 6.64788 4.48206 6.5557 4.25 6.5557C4.01794 6.5557 3.79538 6.64788 3.63128 6.81198C3.46719 6.97607 3.375 7.19863 3.375 7.4307V11.8057C3.375 12.0378 3.46719 12.2603 3.63128 12.4244C3.79538 12.5885 4.01794 12.6807 4.25 12.6807H8.625C8.85706 12.6807 9.07962 12.5885 9.24372 12.4244C9.40781 12.2603 9.5 12.0378 9.5 11.8057C9.5 11.5736 9.40781 11.3511 9.24372 11.187C9.07962 11.0229 8.85706 10.9307 8.625 10.9307H6.10938C6.89141 10.0098 7.67016 9.14242 8.55828 8.24335C9.77437 7.02727 11.3219 6.19652 13.0074 5.85491C14.6929 5.51329 16.4418 5.67597 18.0354 6.3226C19.629 6.96924 20.9967 8.07116 21.9676 9.49067C22.9386 10.9102 23.4697 12.5844 23.4945 14.304C23.5194 16.0236 23.037 17.7124 22.1075 19.1594C21.1781 20.6065 19.8428 21.7475 18.2685 22.4399C16.6943 23.1324 14.9509 23.3456 13.2562 23.0529C11.5615 22.7602 9.9906 21.9745 8.73984 20.7941C8.65625 20.7151 8.55791 20.6534 8.45045 20.6124C8.34299 20.5714 8.2285 20.552 8.11353 20.5552C7.99856 20.5585 7.88536 20.5843 7.78038 20.6313C7.67541 20.6783 7.58072 20.7455 7.50172 20.8291C7.42272 20.9127 7.36096 21.0111 7.31997 21.1185C7.27897 21.226 7.25954 21.3405 7.26279 21.4554C7.26604 21.5704 7.29191 21.6836 7.33891 21.7886C7.38591 21.8936 7.45312 21.9883 7.53672 22.0673C8.78299 23.2433 10.2982 24.0965 11.95 24.5524C13.6018 25.0082 15.3401 25.0528 17.0132 24.6824C18.6863 24.3119 20.2432 23.5376 21.5482 22.427C22.8532 21.3165 23.8665 19.9034 24.4998 18.3111C25.1331 16.7188 25.3671 14.9958 25.1813 13.2923C24.9955 11.5888 24.3956 9.9567 23.434 8.53837C22.4724 7.12005 21.1783 5.95862 19.6646 5.15546C18.1508 4.3523 16.4636 3.93181 14.75 3.9307Z" />
-    </svg>
-  );
-}
-
-export function IconBookmarkFolder({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="28"
-      height="28"
-      viewBox="0 0 28 28"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-      className={cn("fill-[#343330]", className)}
-      {...props}
-    >
-      <path d="M14 22.75C14 22.9821 13.9078 23.2046 13.7437 23.3687C13.5796 23.5328 13.3571 23.625 13.125 23.625H4.375C3.91087 23.625 3.46575 23.4406 3.13756 23.1124C2.80937 22.7842 2.625 22.3391 2.625 21.875V7C2.625 6.53587 2.80937 6.09075 3.13756 5.76256C3.46575 5.43437 3.91087 5.25 4.375 5.25H10.208C10.5865 5.25094 10.9546 5.37365 11.258 5.6L14.292 7.875H23.625C24.0891 7.875 24.5342 8.05938 24.8624 8.38756C25.1906 8.71575 25.375 9.16087 25.375 9.625V13.125C25.375 13.3571 25.2828 13.5796 25.1187 13.7437C24.9546 13.9078 24.7321 14 24.5 14C24.2679 14 24.0454 13.9078 23.8813 13.7437C23.7172 13.5796 23.625 13.3571 23.625 13.125V9.625H14.292C13.9135 9.62406 13.5454 9.50135 13.242 9.275L10.208 7H4.375V21.875H13.125C13.3571 21.875 13.5796 21.9672 13.7437 22.1313C13.9078 22.2954 14 22.5179 14 22.75ZM25.9317 18.3509L23.4697 20.3842L24.2189 23.415C24.26 23.5821 24.251 23.7576 24.1929 23.9196C24.1347 24.0816 24.0302 24.2228 23.8922 24.3257C23.7542 24.4285 23.5889 24.4884 23.4171 24.4978C23.2453 24.5072 23.0745 24.4658 22.9261 24.3786L20.125 22.7303L17.3239 24.3786C17.1755 24.4658 17.0047 24.5072 16.8329 24.4978C16.6611 24.4884 16.4958 24.4285 16.3578 24.3257C16.2198 24.2228 16.1153 24.0816 16.0571 23.9196C15.999 23.7576 15.99 23.5821 16.0311 23.415L16.7803 20.3842L14.3183 18.3509C14.184 18.2399 14.086 18.0913 14.0367 17.9242C13.9874 17.7571 13.9891 17.5791 14.0416 17.4129C14.0942 17.2468 14.1951 17.1001 14.3315 16.9917C14.4678 16.8833 14.6335 16.8181 14.8072 16.8044L18.0709 16.5517L19.3211 13.6533C19.3886 13.4961 19.5008 13.3622 19.6437 13.268C19.7866 13.1739 19.9539 13.1238 20.125 13.1238C20.2961 13.1238 20.4634 13.1739 20.6063 13.268C20.7492 13.3622 20.8614 13.4961 20.9289 13.6533L22.1791 16.5517L25.4428 16.8044C25.6165 16.8181 25.7822 16.8833 25.9185 16.9917C26.0549 17.1001 26.1558 17.2468 26.2084 17.4129C26.2609 17.5791 26.2626 17.7571 26.2133 17.9242C26.164 18.0913 26.066 18.2399 25.9317 18.3509ZM23.1459 18.3816L21.5173 18.2558C21.3585 18.2436 21.206 18.1882 21.0762 18.0957C20.9465 18.0032 20.8445 17.877 20.7812 17.7308L20.125 16.2083L19.4688 17.7308C19.4055 17.877 19.3035 18.0032 19.1738 18.0957C19.044 18.1882 18.8915 18.2436 18.7327 18.2558L17.1041 18.3816L18.3192 19.3856C18.4465 19.4905 18.5416 19.6292 18.5934 19.7858C18.6453 19.9423 18.6518 20.1103 18.6123 20.2705L18.2295 21.8148L19.6809 20.9606C19.8155 20.8814 19.9688 20.8396 20.125 20.8396C20.2812 20.8396 20.4345 20.8814 20.5691 20.9606L22.0205 21.8148L21.6377 20.2705C21.5982 20.1103 21.6047 19.9423 21.6566 19.7858C21.7084 19.6292 21.8035 19.4905 21.9308 19.3856L23.1459 18.3816Z" />
-    </svg>
-  );
-}
-
-export function IconMultiSelect({
-  className,
-  ...props
-}: React.ComponentProps<"svg">) {
-  return (
-    <svg
-      width="24"
-      height="24"
-      viewBox="0 0 24 24"
-      fill="currentColor"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M14.026 8.03497L5.62605 16.285C5.48577 16.423 5.29688 16.5003 5.10011 16.5003C4.90334 16.5003 4.71445 16.423 4.57417 16.285L0.974175 12.7497C0.903938 12.6806 0.847993 12.5984 0.809533 12.5077C0.771074 12.417 0.750853 12.3196 0.750026 12.2211C0.749199 12.1226 0.767782 12.0249 0.804713 11.9336C0.841644 11.8423 0.896201 11.7591 0.965268 11.6889C1.03434 11.6186 1.11656 11.5627 1.20725 11.5242C1.29794 11.4858 1.39531 11.4656 1.49381 11.4647C1.59232 11.4639 1.69002 11.4825 1.78134 11.5194C1.87266 11.5563 1.95581 11.6109 2.02605 11.68L5.10011 14.6987L12.9751 6.96434C13.1171 6.82486 13.3087 6.74748 13.5077 6.74924C13.6062 6.75011 13.7036 6.77038 13.7943 6.8089C13.8851 6.84741 13.9673 6.90342 14.0364 6.97372C14.1054 7.04402 14.16 7.12723 14.1969 7.21861C14.2338 7.30999 14.2523 7.40774 14.2515 7.50628C14.2506 7.60483 14.2303 7.70224 14.1918 7.79295C14.1533 7.88366 14.0973 7.9659 14.027 8.03497H14.026ZM23.0354 6.97372C22.9664 6.90337 22.8842 6.84731 22.7934 6.80876C22.7027 6.77021 22.6053 6.74992 22.5067 6.74905C22.4082 6.74818 22.3104 6.76675 22.219 6.80369C22.1276 6.84063 22.0444 6.89522 21.9742 6.96434L14.0992 14.6987L12.3339 12.9643C12.1919 12.825 12.0004 12.7477 11.8014 12.7496C11.6025 12.7514 11.4124 12.8322 11.2731 12.9742C11.1337 13.1162 11.0565 13.3077 11.0583 13.5066C11.0602 13.7056 11.141 13.8956 11.2829 14.035L13.5732 16.285C13.7135 16.423 13.9024 16.5003 14.0992 16.5003C14.2959 16.5003 14.4848 16.423 14.6251 16.285L23.0251 8.03497C23.0955 7.96599 23.1516 7.88381 23.1903 7.79313C23.2289 7.70245 23.2493 7.60504 23.2502 7.50648C23.2512 7.40792 23.2327 7.31013 23.1959 7.21872C23.159 7.1273 23.1045 7.04405 23.0354 6.97372Z"
-        fill="black"
-      />
-    </svg>
-  );
-}
