+++ src/components/ui/icons/IconArrow.tsx
@@ -1,0 +1,27 @@
+import { cn } from "@/lib/utils";
+
+export default function IconArrow({
+  direction = "up",
+  className,
+  ...props
+}: React.ComponentProps<"svg">) {
+  const paths: Record<string, string> = {
+    up: "M26.0612 15.0613C25.9219 15.2011 25.7563 15.3121 25.574 15.3878C25.3917 15.4635 25.1962 15.5025 24.9987 15.5025C24.8013 15.5025 24.6059 15.4635 24.4235 15.3878C24.2412 15.3121 24.0756 15.2011 23.9362 15.0613L17.5 8.62505V27C17.5 27.3979 17.342 27.7794 17.0607 28.0607C16.7794 28.342 16.3978 28.5 16 28.5C15.6022 28.5 15.2206 28.342 14.9393 28.0607C14.658 27.7794 14.5 27.3979 14.5 27V8.62505L8.06125 15.0613C7.77946 15.3431 7.39726 15.5014 6.99875 15.5014C6.60023 15.5014 6.21804 15.3431 5.93625 15.0613C5.65446 14.7795 5.49615 14.3973 5.49615 13.9988C5.49615 13.6003 5.65446 13.2181 5.93625 12.9363L14.9362 3.9363C15.0756 3.79646 15.2412 3.6855 15.4235 3.6098C15.6058 3.53409 15.8013 3.49512 15.9987 3.49512C16.1962 3.49512 16.3916 3.53409 16.574 3.6098C16.7563 3.6855 16.9219 3.79646 17.0612 3.9363L26.0612 12.9363C26.2011 13.0757 26.312 13.2412 26.3878 13.4236C26.4635 13.6059 26.5024 13.8014 26.5024 13.9988C26.5024 14.1962 26.4635 14.3917 26.3878 14.574C26.312 14.7564 26.2011 14.9219 26.0612 15.0613Z",
+    right:
+      "M17.4622 26.5846C17.3224 26.4453 17.2114 26.2797 17.1357 26.0974C17.06 25.915 17.021 25.7196 17.021 25.5221C17.021 25.3247 17.06 25.1292 17.1357 24.9469C17.2114 24.7646 17.3224 24.599 17.4622 24.4596L23.8984 18.0234H5.52344C5.12562 18.0234 4.74409 17.8653 4.46278 17.584C4.18148 17.3027 4.02344 16.9212 4.02344 16.5234C4.02344 16.1256 4.18148 15.744 4.46278 15.4627C4.74409 15.1814 5.12562 15.0234 5.52344 15.0234L23.8984 15.0234L17.4622 8.58463C17.1804 8.30284 17.0221 7.92065 17.0221 7.52213C17.0221 7.12362 17.1804 6.74143 17.4622 6.45963C17.744 6.17784 18.1262 6.01953 18.5247 6.01953C18.9232 6.01953 19.3054 6.17784 19.5872 6.45963L28.5872 15.4596C28.727 15.599 28.838 15.7646 28.9137 15.9469C28.9894 16.1292 29.0284 16.3247 29.0284 16.5221C29.0284 16.7196 28.9894 16.915 28.9137 17.0974C28.838 17.2797 28.727 17.4453 28.5872 17.5846L19.5872 26.5846C19.4478 26.7245 19.2822 26.8354 19.0999 26.9111C18.9176 26.9868 18.7221 27.0258 18.5247 27.0258C18.3273 27.0258 18.1318 26.9868 17.9495 26.9111C17.7671 26.8354 17.6015 26.7245 17.4622 26.5846Z",
+  };
+
+  return (
+    <svg
+      width="32"
+      height="32"
+      viewBox="0 0 32 32"
+      fill="currentColor"
+      xmlns="http://www.w3.org/2000/svg"
+      className={cn("cursor-pointer", className)}
+      {...props}
+    >
+      <path d={paths[direction]} />
+    </svg>
+  );
+}
