+++ src/components/ui/icons/IconBookmarkFolder.tsx
@@ -1,0 +1,20 @@
+import { cn } from "@/lib/utils";
+
+export default function IconBookmarkFolder({
+  className,
+  ...props
+}: React.ComponentProps<"svg">) {
+  return (
+    <svg
+      width="28"
+      height="28"
+      viewBox="0 0 28 28"
+      fill="currentColor"
+      xmlns="http://www.w3.org/2000/svg"
+      className={cn("fill-[#343330]", className)}
+      {...props}
+    >
+      <path d="M14 22.75C14 22.9821 13.9078 23.2046 13.7437 23.3687C13.5796 23.5328 13.3571 23.625 13.125 23.625H4.375C3.91087 23.625 3.46575 23.4406 3.13756 23.1124C2.80937 22.7842 2.625 22.3391 2.625 21.875V7C2.625 6.53587 2.80937 6.09075 3.13756 5.76256C3.46575 5.43437 3.91087 5.25 4.375 5.25H10.208C10.5865 5.25094 10.9546 5.37365 11.258 5.6L14.292 7.875H23.625C24.0891 7.875 24.5342 8.05938 24.8624 8.38756C25.1906 8.71575 25.375 9.16087 25.375 9.625V13.125C25.375 13.3571 25.2828 13.5796 25.1187 13.7437C24.9546 13.9078 24.7321 14 24.5 14C24.2679 14 24.0454 13.9078 23.8813 13.7437C23.7172 13.5796 23.625 13.3571 23.625 13.125V9.625H14.292C13.9135 9.62406 13.5454 9.50135 13.242 9.275L10.208 7H4.375V21.875H13.125C13.3571 21.875 13.5796 21.9672 13.7437 22.1313C13.9078 22.2954 14 22.5179 14 22.75ZM25.9317 18.3509L23.4697 20.3842L24.2189 23.415C24.26 23.5821 24.251 23.7576 24.1929 23.9196C24.1347 24.0816 24.0302 24.2228 23.8922 24.3257C23.7542 24.4285 23.5889 24.4884 23.4171 24.4978C23.2453 24.5072 23.0745 24.4658 22.9261 24.3786L20.125 22.7303L17.3239 24.3786C17.1755 24.4658 17.0047 24.5072 16.8329 24.4978C16.6611 24.4884 16.4958 24.4285 16.3578 24.3257C16.2198 24.2228 16.1153 24.0816 16.0571 23.9196C15.999 23.7576 15.99 23.5821 16.0311 23.415L16.7803 20.3842L14.3183 18.3509C14.184 18.2399 14.086 18.0913 14.0367 17.9242C13.9874 17.7571 13.9891 17.5791 14.0416 17.4129C14.0942 17.2468 14.1951 17.1001 14.3315 16.9917C14.4678 16.8833 14.6335 16.8181 14.8072 16.8044L18.0709 16.5517L19.3211 13.6533C19.3886 13.4961 19.5008 13.3622 19.6437 13.268C19.7866 13.1739 19.9539 13.1238 20.125 13.1238C20.2961 13.1238 20.4634 13.1739 20.6063 13.268C20.7492 13.3622 20.8614 13.4961 20.9289 13.6533L22.1791 16.5517L25.4428 16.8044C25.6165 16.8181 25.7822 16.8833 25.9185 16.9917C26.0549 17.1001 26.1558 17.2468 26.2084 17.4129C26.2609 17.5791 26.2626 17.7571 26.2133 17.9242C26.164 18.0913 26.066 18.2399 25.9317 18.3509ZM23.1459 18.3816L21.5173 18.2558C21.3585 18.2436 21.206 18.1882 21.0762 18.0957C20.9465 18.0032 20.8445 17.877 20.7812 17.7308L20.125 16.2083L19.4688 17.7308C19.4055 17.877 19.3035 18.0032 19.1738 18.0957C19.044 18.1882 18.8915 18.2436 18.7327 18.2558L17.1041 18.3816L18.3192 19.3856C18.4465 19.4905 18.5416 19.6292 18.5934 19.7858C18.6453 19.9423 18.6518 20.1103 18.6123 20.2705L18.2295 21.8148L19.6809 20.9606C19.8155 20.8814 19.9688 20.8396 20.125 20.8396C20.2812 20.8396 20.4345 20.8814 20.5691 20.9606L22.0205 21.8148L21.6377 20.2705C21.5982 20.1103 21.6047 19.9423 21.6566 19.7858C21.7084 19.6292 21.8035 19.4905 21.9308 19.3856L23.1459 18.3816Z" />
+    </svg>
+  );
+}
