+++ src/components/ui/icons/IconBug.tsx
@@ -1,0 +1,62 @@
+import { cn } from "@/lib/utils";
+
+export default function IconBug({
+  className,
+  width = "191",
+  height = "196",
+  color = "fill-primary-500",
+  ...props
+}: React.ComponentProps<"svg"> & { width?: string; height?: string }) {
+  return (
+    <svg
+      width={width}
+      height={height}
+      viewBox="0 0 191 196"
+      fill="currentColor"
+      xmlns="http://www.w3.org/2000/svg"
+      className={cn(color, className)}
+      {...props}
+    >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M95.3023 20.0367C69.6385 20.0367 50.4865 38.8569 50.4865 60.1101C50.4865 65.6431 46.0277 70.1284 40.5274 70.1284C35.0272 70.1284 30.5684 65.6431 30.5684 60.1101C30.5684 26.0366 60.467 0 95.3023 0C130.138 0 160.036 26.0366 160.036 60.1101C160.036 65.6431 155.577 70.1284 150.077 70.1284C144.577 70.1284 140.118 65.6431 140.118 60.1101C140.118 38.8569 120.966 20.0367 95.3023 20.0367Z"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M40.5274 120.477C46.0277 120.477 50.4865 124.962 50.4865 130.495C50.4865 151.749 69.6385 170.569 95.3023 170.569C120.966 170.569 140.118 151.749 140.118 130.495C140.118 124.962 144.577 120.477 150.077 120.477C155.577 120.477 160.036 124.962 160.036 130.495C160.036 164.569 130.138 190.605 95.3023 190.605C60.467 190.605 30.5684 164.569 30.5684 130.495C30.5684 124.962 35.0272 120.477 40.5274 120.477Z"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M0 96.2024C0 90.7404 4.49142 86.3125 10.0319 86.3125H180.574C186.114 86.3125 190.606 90.7404 190.606 96.2024C190.606 101.664 186.114 106.092 180.574 106.092H10.0319C4.49142 106.092 0 101.664 0 96.2024Z"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M96.2024 131.267C101.664 131.267 106.092 135.658 106.092 141.075V175.404C106.092 180.821 101.664 185.212 96.2024 185.212C90.7404 185.212 86.3125 180.821 86.3125 175.404V141.075C86.3125 135.658 90.7404 131.267 96.2024 131.267Z"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M33.4274 3.00955C37.2395 -1.00318 43.4201 -1.00318 47.2322 3.00955L61.8744 18.4224C65.6865 22.4351 65.6865 28.941 61.8744 32.9538C58.0623 36.9665 51.8817 36.9665 48.0696 32.9538L33.4274 17.5409C29.6153 13.5282 29.6153 7.02227 33.4274 3.00955Z"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M157.178 3.00955C160.99 7.02227 160.99 13.5282 157.178 17.5409L142.536 32.9538C138.724 36.9665 132.543 36.9665 128.731 32.9538C124.919 28.941 124.919 22.4351 128.731 18.4224L143.373 3.00955C147.185 -1.00318 153.366 -1.00318 157.178 3.00955Z"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M61.8744 163.047C65.6865 167.059 65.6865 173.565 61.8744 177.578L47.2322 192.991C43.4201 197.004 37.2395 197.004 33.4274 192.991C29.6153 188.978 29.6153 182.472 33.4274 178.459L48.0696 163.047C51.8817 159.034 58.0623 159.034 61.8744 163.047Z"
+      />
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M128.731 163.047C132.543 159.034 138.724 159.034 142.536 163.047L157.178 178.459C160.99 182.472 160.99 188.978 157.178 192.991C153.366 197.004 147.185 197.004 143.373 192.991L128.731 177.578C124.919 173.565 124.919 167.059 128.731 163.047Z"
+      />
+    </svg>
+  );
+}
