+++ src/components/ui/icons/IconCaretDoubleDown.tsx
@@ -1,0 +1,23 @@
+import { cn } from "@/lib/utils";
+
+export default function IconCaretDoubleDown({
+  className,
+  ...props
+}: React.ComponentProps<"svg">) {
+  return (
+    <svg
+      width="56"
+      height="57"
+      viewBox="0 0 56 57"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={cn("fill-primary-500", className)}
+      {...props}
+    >
+      <path
+        d="M46.7381 27.2619C46.9008 27.4244 47.0299 27.6174 47.118 27.8299C47.206 28.0423 47.2514 28.27 47.2514 28.5C47.2514 28.73 47.206 28.9577 47.118 29.1701C47.0299 29.3826 46.9008 29.5756 46.7381 29.7381L29.2381 47.2381C29.0756 47.4008 28.8826 47.5299 28.6701 47.618C28.4577 47.706 28.23 47.7514 28 47.7514C27.77 47.7514 27.5423 47.706 27.3299 47.618C27.1174 47.5299 26.9244 47.4008 26.7619 47.2381L9.26187 29.7381C8.9335 29.4098 8.74902 28.9644 8.74902 28.5C8.74902 28.0356 8.9335 27.5902 9.26187 27.2619C9.59024 26.9335 10.0356 26.749 10.5 26.749C10.9644 26.749 11.4098 26.9335 11.7381 27.2619L28 43.5259L44.2619 27.2619C44.4244 27.0992 44.6174 26.9701 44.8298 26.882C45.0423 26.7939 45.27 26.7486 45.5 26.7486C45.73 26.7486 45.9577 26.7939 46.1701 26.882C46.3826 26.9701 46.5756 27.0992 46.7381 27.2619ZM26.7619 29.7381C26.9244 29.9008 27.1174 30.0299 27.3299 30.118C27.5423 30.206 27.77 30.2514 28 30.2514C28.23 30.2514 28.4577 30.206 28.6701 30.118C28.8826 30.0299 29.0756 29.9008 29.2381 29.7381L46.7381 12.2381C47.0665 11.9098 47.251 11.4644 47.251 11C47.251 10.5356 47.0665 10.0902 46.7381 9.76187C46.4098 9.4335 45.9644 9.24902 45.5 9.24902C45.0356 9.24902 44.5902 9.4335 44.2619 9.76187L28 26.0259L11.7381 9.76187C11.4098 9.4335 10.9644 9.24902 10.5 9.24902C10.0356 9.24902 9.59024 9.4335 9.26187 9.76187C8.9335 10.0902 8.74902 10.5356 8.74902 11C8.74902 11.4644 8.9335 11.9098 9.26187 12.2381L26.7619 29.7381Z"
+        fill="#6100FF"
+      />
+    </svg>
+  );
+}
