+++ src/components/ui/icons/IconChat.tsx
@@ -1,0 +1,23 @@
+import { cn } from "@/lib/utils";
+
+export default function IconChat({
+  className,
+  ...props
+}: React.ComponentProps<"svg">) {
+  return (
+    <svg
+      width="35"
+      height="35"
+      viewBox="0 0 35 35"
+      fill="currentColor"
+      xmlns="http://www.w3.org/2000/svg"
+      className={cn(className)}
+      {...props}
+    >
+      <path
+        d="M33.8906 19.6112C32.8406 27.6437 24.8956 34.1537 16.7931 34.1537H5.12059C4.12986 34.1541 3.21173 33.6342 2.70253 32.7843C2.19334 31.9345 2.16792 30.8796 2.63559 30.0062L3.10809 29.1137C3.61559 28.2494 3.61559 27.1781 3.10809 26.3137C-0.86958 20.0304 -0.248309 11.8807 4.63584 6.27293C9.51999 0.665183 17.5073 -1.06902 24.2772 2.00841C31.047 5.08583 34.992 12.2441 33.9781 19.6112H33.8906Z"
+        fill="currentColor"
+      />
+    </svg>
+  );
+}
