+++ src/components/ui/icons/IconClockCounterClockwise.tsx
@@ -1,0 +1,20 @@
+import { cn } from "@/lib/utils";
+
+export default function IconClockCounterClockwise({
+  className,
+  ...props
+}: React.ComponentProps<"svg">) {
+  return (
+    <svg
+      width="29"
+      height="29"
+      viewBox="0 0 29 29"
+      fill="currentColor"
+      xmlns="http://www.w3.org/2000/svg"
+      className={cn("fill-[#343330]", className)}
+      {...props}
+    >
+      <path d="M15.625 9.1807V13.9352L19.5756 16.3054C19.7746 16.4249 19.918 16.6186 19.9742 16.8438C20.0304 17.069 19.9948 17.3073 19.8753 17.5063C19.7558 17.7053 19.5621 17.8487 19.3369 17.9049C19.1117 17.9611 18.8734 17.9255 18.6744 17.806L14.2994 15.181C14.1699 15.1032 14.0627 14.9932 13.9884 14.8617C13.914 14.7303 13.875 14.5818 13.875 14.4307V9.1807C13.875 8.94863 13.9672 8.72607 14.1313 8.56198C14.2954 8.39788 14.5179 8.3057 14.75 8.3057C14.9821 8.3057 15.2046 8.39788 15.3687 8.56198C15.5328 8.72607 15.625 8.94863 15.625 9.1807ZM14.75 3.9307C13.3697 3.92726 12.0024 4.19761 10.7272 4.7261C9.45211 5.25459 8.29444 6.03073 7.32125 7.0096C6.52609 7.8146 5.81953 8.58898 5.125 9.39945V7.4307C5.125 7.19863 5.03281 6.97607 4.86872 6.81198C4.70462 6.64788 4.48206 6.5557 4.25 6.5557C4.01794 6.5557 3.79538 6.64788 3.63128 6.81198C3.46719 6.97607 3.375 7.19863 3.375 7.4307V11.8057C3.375 12.0378 3.46719 12.2603 3.63128 12.4244C3.79538 12.5885 4.01794 12.6807 4.25 12.6807H8.625C8.85706 12.6807 9.07962 12.5885 9.24372 12.4244C9.40781 12.2603 9.5 12.0378 9.5 11.8057C9.5 11.5736 9.40781 11.3511 9.24372 11.187C9.07962 11.0229 8.85706 10.9307 8.625 10.9307H6.10938C6.89141 10.0098 7.67016 9.14242 8.55828 8.24335C9.77437 7.02727 11.3219 6.19652 13.0074 5.85491C14.6929 5.51329 16.4418 5.67597 18.0354 6.3226C19.629 6.96924 20.9967 8.07116 21.9676 9.49067C22.9386 10.9102 23.4697 12.5844 23.4945 14.304C23.5194 16.0236 23.037 17.7124 22.1075 19.1594C21.1781 20.6065 19.8428 21.7475 18.2685 22.4399C16.6943 23.1324 14.9509 23.3456 13.2562 23.0529C11.5615 22.7602 9.9906 21.9745 8.73984 20.7941C8.65625 20.7151 8.55791 20.6534 8.45045 20.6124C8.34299 20.5714 8.2285 20.552 8.11353 20.5552C7.99856 20.5585 7.88536 20.5843 7.78038 20.6313C7.67541 20.6783 7.58072 20.7455 7.50172 20.8291C7.42272 20.9127 7.36096 21.0111 7.31997 21.1185C7.27897 21.226 7.25954 21.3405 7.26279 21.4554C7.26604 21.5704 7.29191 21.6836 7.33891 21.7886C7.38591 21.8936 7.45312 21.9883 7.53672 22.0673C8.78299 23.2433 10.2982 24.0965 11.95 24.5524C13.6018 25.0082 15.3401 25.0528 17.0132 24.6824C18.6863 24.3119 20.2432 23.5376 21.5482 22.427C22.8532 21.3165 23.8665 19.9034 24.4998 18.3111C25.1331 16.7188 25.3671 14.9958 25.1813 13.2923C24.9955 11.5888 24.3956 9.9567 23.434 8.53837C22.4724 7.12005 21.1783 5.95862 19.6646 5.15546C18.1508 4.3523 16.4636 3.93181 14.75 3.9307Z" />
+    </svg>
+  );
+}
