+++ src/components/ui/icons/IconClose.tsx
@@ -1,0 +1,20 @@
+import { cn } from "@/lib/utils";
+
+export default function IconClose({
+  className,
+  ...props
+}: React.ComponentProps<"svg">) {
+  return (
+    <svg
+      width="12"
+      height="12"
+      viewBox="0 0 12 12"
+      fill="currentColor"
+      xmlns="http://www.w3.org/2000/svg"
+      className={cn(className)}
+      {...props}
+    >
+      <path d="M7.30975 5.99999L9.78007 3.53061C9.84975 3.46093 9.90502 3.37821 9.94273 3.28716C9.98045 3.19612 9.99986 3.09854 9.99986 2.99999C9.99986 2.90144 9.98045 2.80386 9.94273 2.71282C9.90502 2.62177 9.84975 2.53905 9.78007 2.46936C9.71038 2.39968 9.62766 2.34441 9.53661 2.30669C9.44557 2.26898 9.34799 2.24957 9.24944 2.24957C9.15089 2.24957 9.05331 2.26898 8.96227 2.30669C8.87122 2.34441 8.7885 2.39968 8.71881 2.46936L6.24944 4.93968L3.78007 2.46936C3.63934 2.32863 3.44846 2.24957 3.24944 2.24957C3.05042 2.24957 2.85955 2.32863 2.71882 2.46936C2.57808 2.61009 2.49902 2.80097 2.49902 2.99999C2.49902 3.19901 2.57808 3.38988 2.71882 3.53061L5.18913 5.99999L2.71882 8.46936C2.64913 8.53905 2.59386 8.62177 2.55615 8.71282C2.51843 8.80386 2.49902 8.90144 2.49902 8.99999C2.49902 9.09854 2.51843 9.19612 2.55615 9.28716C2.59386 9.37821 2.64913 9.46093 2.71882 9.53061C2.7885 9.6003 2.87122 9.65557 2.96227 9.69328C3.05331 9.731 3.15089 9.75041 3.24944 9.75041C3.34799 9.75041 3.44557 9.731 3.53661 9.69328C3.62766 9.65557 3.71038 9.6003 3.78007 9.53061L6.24944 7.0603L8.71881 9.53061C8.7885 9.6003 8.87122 9.65557 8.96227 9.69328C9.05331 9.731 9.15089 9.75041 9.24944 9.75041C9.34799 9.75041 9.44557 9.731 9.53661 9.69328C9.62766 9.65557 9.71038 9.6003 9.78007 9.53061C9.84975 9.46093 9.90502 9.37821 9.94273 9.28716C9.98045 9.19612 9.99986 9.09854 9.99986 8.99999C9.99986 8.90144 9.98045 8.80386 9.94273 8.71282C9.90502 8.62177 9.84975 8.53905 9.78007 8.46936L7.30975 5.99999Z" />
+    </svg>
+  );
+}
