+++ src/components/ui/icons/IconError.tsx
@@ -1,0 +1,25 @@
+import { cn } from "@/lib/utils";
+
+export default function IconError({
+  className,
+  ...props
+}: React.ComponentProps<"svg">) {
+  return (
+    <svg
+      width="20"
+      height="20"
+      viewBox="0 0 20 20"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={cn(className)}
+      {...props}
+    >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M17.3796 12.9196L12.6796 4.21959C12.0926 3.15854 10.9756 2.5 9.76298 2.5C8.55035 2.5 7.43337 3.15854 6.84631 4.21959L2.07131 12.9113C1.51031 13.9435 1.53416 15.1949 2.1341 16.2051C2.73403 17.2152 3.82144 17.8349 4.99631 17.8363H14.4463C15.6202 17.8363 16.7075 17.2189 17.3089 16.2108C17.9104 15.2027 17.9372 13.9526 17.3796 12.9196ZM9.04631 8.11959C9.04631 7.77441 9.32613 7.49459 9.67131 7.49459C10.0165 7.49459 10.2963 7.77441 10.2963 8.11959V10.6946C10.2963 11.0398 10.0165 11.3196 9.67131 11.3196C9.32613 11.3196 9.04631 11.0398 9.04631 10.6946V8.11959ZM9.67964 13.7446C10.0248 13.7446 10.3046 13.4648 10.3046 13.1196L10.2963 13.1113C10.2963 12.7707 10.0202 12.4946 9.67964 12.4946C9.33634 12.4991 9.05913 12.7763 9.05464 13.1196C9.05464 13.4648 9.33447 13.7446 9.67964 13.7446Z"
+        fill="#FF6D6D"
+      />
+    </svg>
+  );
+}
