+++ src/components/ui/icons/IconFolder.tsx
@@ -1,0 +1,24 @@
+import { cn } from "@/lib/utils";
+
+export default function IconFolder({
+  className,
+  ...props
+}: React.ComponentProps<"svg">) {
+  return (
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      fill="currentColor"
+      xmlns="http://www.w3.org/2000/svg"
+      className={cn(className)}
+      {...props}
+    >
+      <path
+        fillRule="evenodd"
+        clipRule="evenodd"
+        d="M19.45 7.625V7.905C19.9453 8.52698 20.146 9.33343 20 10.115L17.92 16.615C17.5115 17.6736 16.4947 18.3727 15.36 18.375H2.7C1.21076 18.3427 0.015841 17.1345 1.51304e-08 15.645V4.375C-0.000182366 2.3233 1.64848 0.652356 3.7 0.625H8C9.38071 0.625 10.5 1.74429 10.5 3.125C10.5 3.67728 10.9477 4.125 11.5 4.125H16C17.9134 4.15234 19.4502 5.71136 19.45 7.625ZM3.7 2.125C2.45736 2.125 1.45 3.13236 1.45 4.375V11.495L2.52 8.715C2.91802 7.63378 3.94784 6.91529 5.1 6.915H17.35C17.523 6.89995 17.697 6.89995 17.87 6.915C17.5751 6.13835 16.8308 5.62489 16 5.625H11.5C10.1216 5.61952 9.00548 4.50344 9 3.125C9 2.57272 8.55229 2.125 8 2.125H3.7Z"
+      />
+    </svg>
+  );
+}
