+++ src/components/ui/icons/IconHourGlass.tsx
@@ -1,0 +1,20 @@
+import { cn } from "@/lib/utils";
+
+export default function IconHourGlass({
+  className,
+  ...props
+}: React.ComponentProps<"svg">) {
+  return (
+    <svg
+      width="48"
+      height="48"
+      viewBox="0 0 48 48"
+      fill="currentColor"
+      xmlns="http://www.w3.org/2000/svg"
+      className={cn("fill-gray-dark", className)}
+      {...props}
+    >
+      <path d="M37.5 14.1825V7.5C37.5 6.70435 37.1839 5.94129 36.6213 5.37868C36.0587 4.81607 35.2956 4.5 34.5 4.5H13.5C12.7044 4.5 11.9413 4.81607 11.3787 5.37868C10.8161 5.94129 10.5 6.70435 10.5 7.5V14.25C10.501 14.7156 10.6099 15.1745 10.8181 15.591C11.0263 16.0074 11.3282 16.3699 11.7 16.65L21.5006 24L11.7 31.35C11.3282 31.6301 11.0263 31.9926 10.8181 32.409C10.6099 32.8255 10.501 33.2844 10.5 33.75V40.5C10.5 41.2956 10.8161 42.0587 11.3787 42.6213C11.9413 43.1839 12.7044 43.5 13.5 43.5H34.5C35.2956 43.5 36.0587 43.1839 36.6213 42.6213C37.1839 42.0587 37.5 41.2956 37.5 40.5V33.8175C37.4989 33.3537 37.391 32.8965 37.1845 32.4812C36.978 32.0659 36.6785 31.7039 36.3094 31.4231L26.4881 24L36.3094 16.5769C36.6785 16.2961 36.978 15.9341 37.1845 15.5188C37.391 15.1035 37.4989 14.6463 37.5 14.1825ZM34.5 7.5V12H13.5V7.5H34.5ZM34.5 40.5H13.5V33.75L24 25.875L34.5 33.8156V40.5Z" />
+    </svg>
+  );
+}
