+++ src/components/ui/icons/IconMagnifierWithPlus.tsx
@@ -1,0 +1,23 @@
+import { cn } from "@/lib/utils";
+
+export default function IconMagnifierWithPlus({
+  className,
+  ...props
+}: React.ComponentProps<"svg">) {
+  return (
+    <svg
+      width="27"
+      height="27"
+      viewBox="0 0 27 27"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={cn("fill-primary-500", className)}
+      {...props}
+    >
+      <path
+        d="M26.2078 25.2923L19.949 19.0348C21.7631 16.8569 22.6676 14.0635 22.4746 11.2357C22.2815 8.40789 21.0057 5.76337 18.9125 3.85226C16.8193 1.94116 14.0698 0.910614 11.2362 0.975014C8.4025 1.03941 5.70274 2.1938 3.69851 4.19802C1.69429 6.20225 0.539902 8.90201 0.475503 11.7357C0.411103 14.5694 1.44164 17.3188 3.35275 19.412C5.26385 21.5052 7.90838 22.781 10.7362 22.9741C13.564 23.1672 16.3574 22.2626 18.5353 20.4485L24.7928 26.7073C24.8857 26.8002 24.996 26.8739 25.1174 26.9242C25.2388 26.9745 25.3689 27.0004 25.5003 27.0004C25.6317 27.0004 25.7618 26.9745 25.8832 26.9242C26.0046 26.8739 26.1149 26.8002 26.2078 26.7073C26.3007 26.6144 26.3744 26.5041 26.4247 26.3827C26.475 26.2613 26.5008 26.1312 26.5008 25.9998C26.5008 25.8684 26.475 25.7383 26.4247 25.6169C26.3744 25.4955 26.3007 25.3852 26.2078 25.2923ZM2.50029 11.9998C2.50029 10.2198 3.02813 8.47971 4.01706 6.99966C5.00599 5.51962 6.4116 4.36607 8.05613 3.68488C9.70067 3.00369 11.5103 2.82546 13.2561 3.17273C15.0019 3.52 16.6056 4.37716 17.8642 5.63584C19.1229 6.89451 19.9801 8.49816 20.3274 10.244C20.6746 11.9898 20.4964 13.7994 19.8152 15.4439C19.134 17.0885 17.9805 18.4941 16.5004 19.483C15.0204 20.472 13.2803 20.9998 11.5003 20.9998C9.11415 20.9972 6.8265 20.0481 5.13925 18.3608C3.45199 16.6736 2.50293 14.3859 2.50029 11.9998Z"
+        fill="#6100FF"
+      />
+    </svg>
+  );
+}
