+++ src/components/ui/icons/IconOpenFolder.tsx
@@ -1,0 +1,18 @@
+export default function IconOpenFolder({
+  className,
+  ...props
+}: React.ComponentProps<"svg">) {
+  return (
+    <svg
+      width="13"
+      height="13"
+      viewBox="0 0 13 13"
+      fill="currentColor"
+      xmlns="http://www.w3.org/2000/svg"
+      className={className}
+      {...props}
+    >
+      <path d="M12.3646 5.73855C12.2932 5.63955 12.1993 5.55896 12.0906 5.50343C11.9819 5.44791 11.8615 5.41904 11.7394 5.41922H10.97V4.64975C10.97 4.44567 10.8889 4.24996 10.7446 4.10565C10.6003 3.96135 10.4046 3.88028 10.2005 3.88028H6.86629L5.5327 2.87997C5.39928 2.78051 5.23743 2.72656 5.07102 2.72607H2.5058C2.30172 2.72607 2.106 2.80714 1.9617 2.95145C1.8174 3.09575 1.73633 3.29147 1.73633 3.49554V10.4208C1.73633 10.5228 1.77686 10.6207 1.84901 10.6928C1.92117 10.765 2.01903 10.8055 2.12106 10.8055H10.7343C10.8151 10.8055 10.8938 10.7801 10.9593 10.7329C11.0248 10.6857 11.0738 10.6191 11.0993 10.5424L12.4695 6.43204C12.508 6.31638 12.5186 6.19324 12.5004 6.0727C12.4821 5.95216 12.4356 5.83765 12.3646 5.73855ZM5.07102 3.49554L6.40461 4.49586C6.53799 4.59538 6.69987 4.64934 6.86629 4.64975H10.2005V5.41922H3.93749C3.77599 5.41921 3.61857 5.47001 3.48755 5.56444C3.35653 5.65887 3.25853 5.79213 3.20746 5.94535L2.5058 8.04985V3.49554H5.07102Z" />
+    </svg>
+  );
+}
