+++ src/components/ui/icons/IconShare.tsx
@@ -1,0 +1,27 @@
+import { cn } from "@/lib/utils";
+
+export default function IconShare({
+  className,
+  ...props
+}: React.ComponentProps<"svg">) {
+  return (
+    <svg
+      width="28"
+      height="23"
+      viewBox="0 0 28 23"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+      className={cn(className)}
+      {...props}
+    >
+      <path
+        d="M27.7074 10.2933L17.7074 0.293287C17.5676 0.153351 17.3895 0.0580095 17.1955 0.0193193C17.0016 -0.0193709 16.8005 0.000328541 16.6177 0.075926C16.4349 0.151524 16.2787 0.279623 16.1687 0.444026C16.0588 0.608428 16 0.801748 15.9999 0.999537V6.04329C12.7574 6.32079 9.17617 7.90829 6.22992 10.407C2.68242 13.417 0.473674 17.2958 0.00992401 21.3283C-0.0263163 21.6418 0.0372509 21.9587 0.191579 22.234C0.345907 22.5093 0.583132 22.7289 0.869495 22.8616C1.15586 22.9942 1.47677 23.0332 1.78655 22.9729C2.09634 22.9126 2.37921 22.7562 2.59492 22.5258C3.96992 21.062 8.86242 16.4333 15.9999 16.0258V20.9995C16 21.1973 16.0588 21.3906 16.1687 21.555C16.2787 21.7194 16.4349 21.8475 16.6177 21.9231C16.8005 21.9987 17.0016 22.0184 17.1955 21.9798C17.3895 21.9411 17.5676 21.8457 17.7074 21.7058L27.7074 11.7058C27.8944 11.5183 27.9995 11.2643 27.9995 10.9995C27.9995 10.7347 27.8944 10.4808 27.7074 10.2933ZM17.9999 18.5858V14.9995C17.9999 14.7343 17.8946 14.48 17.707 14.2924C17.5195 14.1049 17.2651 13.9995 16.9999 13.9995C13.4899 13.9995 10.0712 14.9158 6.83867 16.7245C5.19235 17.6498 3.65844 18.7622 2.26742 20.0395C2.99242 17.0595 4.81992 14.2258 7.52367 11.932C10.4262 9.47079 13.9687 7.99954 16.9999 7.99954C17.2651 7.99954 17.5195 7.89418 17.707 7.70664C17.8946 7.51911 17.9999 7.26475 17.9999 6.99954V3.41454L25.5862 10.9995L17.9999 18.5858Z"
+        fill="#6100FF"
+      />
+      <path
+        d="M17.9999 18.5858V14.9995C17.9999 14.7343 17.8946 14.48 17.707 14.2924C17.5195 14.1049 17.2651 13.9995 16.9999 13.9995C13.4899 13.9995 10.0712 14.9158 6.83867 16.7245C5.19235 17.6498 3.65844 18.7622 2.26742 20.0395C2.99242 17.0595 4.81992 14.2258 7.52367 11.932C10.4262 9.47079 13.9687 7.99954 16.9999 7.99954C17.2651 7.99954 17.5195 7.89418 17.707 7.70664C17.8946 7.51911 17.9999 7.26475 17.9999 6.99954V3.41454L25.5862 10.9995L17.9999 18.5858Z"
+        fill="#FFFFFF"
+      />
+    </svg>
+  );
+}
