+++ src/components/ui/icons/IconStar.tsx
@@ -1,0 +1,25 @@
+import { cn } from "@/lib/utils";
+
+export default function IconStar({
+  className,
+  filled = false,
+  ...props
+}: React.ComponentProps<"svg"> & { filled?: boolean }) {
+  return (
+    <svg
+      width="24"
+      height="24"
+      viewBox="0 0 24 24"
+      className={cn(
+        "transition-colors",
+        filled ? "fill-current" : "fill-none stroke-current",
+        className,
+      )}
+      strokeWidth={filled ? "0" : "1.5"}
+      xmlns="http://www.w3.org/2000/svg"
+      {...props}
+    >
+      <path d="M22.4258 9.12124C22.335 8.84129 22.1638 8.59424 21.9336 8.41087C21.7034 8.2275 21.4243 8.11591 21.1311 8.08999L15.5633 7.60999L13.3808 2.41999C13.2672 2.14766 13.0755 1.91504 12.8299 1.75142C12.5844 1.5878 12.2959 1.50049 12.0008 1.50049C11.7057 1.50049 11.4172 1.5878 11.1717 1.75142C10.9261 1.91504 10.7344 2.14766 10.6208 2.41999L8.44486 7.60999L2.87049 8.0928C2.57615 8.11754 2.29566 8.22862 2.0642 8.41213C1.83274 8.59564 1.66061 8.8434 1.5694 9.12434C1.47819 9.40528 1.47195 9.7069 1.55148 9.99138C1.631 10.2758 1.79275 10.5305 2.01643 10.7234L6.24549 14.4191L4.97799 19.9072C4.91097 20.1944 4.93009 20.495 5.03297 20.7713C5.13585 21.0477 5.31792 21.2877 5.55642 21.4611C5.79492 21.6346 6.07926 21.7339 6.3739 21.7467C6.66854 21.7594 6.9604 21.685 7.21299 21.5328L11.9942 18.6266L16.7858 21.5328C17.0384 21.685 17.3303 21.7594 17.6249 21.7467C17.9195 21.7339 18.2039 21.6346 18.4424 21.4611C18.6809 21.2877 18.8629 21.0477 18.9658 20.7713C19.0687 20.495 19.0878 20.1944 19.0208 19.9072L17.7542 14.4134L21.9824 10.7234C22.206 10.5298 22.3674 10.2745 22.4463 9.98938C22.5252 9.70431 22.5181 9.40227 22.4258 9.12124Z" />
+    </svg>
+  );
+}
