+++ src/components/vulnerability-db/Search.tsx
@@ -2,7 +2,7 @@
 
 import { updateRealTimeTopic } from "@/lib/api/posts";
 import React, { useEffect, useState } from "react";
-import { IconMagnifierWithPlus } from "../ui/Icons";
+import IconMagnifierWithPlus from "../ui/icons/IconMagnifierWithPlus";
 import { showToast } from "./Toast";
 import { processSearchInput } from "@/lib/utils";
