+++ src/components/vulnerability-db/VulDBPin.tsx
@@ -5,7 +5,8 @@ import { addPinnedPostToUser, deletePinnedPostFromUser } from "@/lib/api/users";
 import { VulDBPinnedInfo } from "@/types/post";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useState } from "react";
-import { IconFilledPin, IconPin } from "../ui/Icons";
+import IconFilledPin from "../ui/icons/IconFilledPin";
+import IconPin from "../ui/icons/IconPin";
 
 export default function VulDBPin({
   pinnedInfo,
