+++ src/components/vulnerability-db/VulDBShare.tsx
@@ -1,8 +1,9 @@
 "use client";
 
-import { IconFilledShare, IconShare } from "@/components/ui/Icons";
 import { useState } from "react";
 import { showToast } from "./Toast";
+import IconShare from "../ui/icons/IconShare";
+import IconFilledShare from "../ui/icons/IconFilledShare";
 
 export default function VulDBShare({ postId }: { postId: string }) {
   const [isCopied, setIsCopied] = useState(false);
